### 2025-01-21 #233 @syedalimohsinbukhari
`pymultifit` submission
Submitting Author: Syed Ali Mohsin Bukhari (@syedalimohsinbukhari)
Package Name: `pymultifit`
One-Line Description of Package: A python library for fitting data with multiple models.
Repository Link: https://github.com/syedalimohsinbukhari/pyMultiFit
Version submitted: ~v1.0.3~ [v1.0.6](https://github.com/syedalimohsinbukhari/pyMultiFit/releases/tag/v1.0.6)
Archive: https://zenodo.org/records/17273507
JOSS DOI: https://doi.org/10.21105/joss.09295
Version accepted: 1.0.7

---

`pymultifit` is built primarily to solve one problem, to fit multiple models (and mixture models) to a given data. Be it multiple `Gaussian`, `Laplacian`, or a mixture of such models, this package aims to deal with multi-model data fitting. The package also provides easy-to-use `BaseDistribution` and `BaseFitter` classes for respective user-defined functions.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Researchers, data scientists, and statisticians who work with datasets requiring multi-model fitting for robust analysis and modeling.

Apart from scipy, lmfit, and scikit-learn the general purpose scientific packages, there exists [PyAutoFit](https://github.com/rhayes777/PyAutoFit), a Python-based probabilistic programming language built on Bayesian inference. Another notable library is [Mixture-Models](https://github.com/kasakh/Mixture-Models), which specializes in advanced optimization techniques for fitting various families of mixture models, including Gaussian mixture models and their variants. Both libraries are powerful tools for specific use cases, and I recently came to know about them during my search of existing options.

While these libraries offer robust solutions for hierarchical modeling (PyAutoFit) or a diverse array of pre-defined mixture models (Mixture-Models), pyMultiFit distinguishes itself through its simplicity of use and its focus on simplicity of use. Specifically, it is designed to provide a lightweight and user-friendly framework for fitting multi-model data, including custom mixture models (for example, gaussian + laplace + line). pymultifit also provides easy-to-use base classes that can be modified for any distribution/fitter purposes.

One of the more prominent features of pyMultiFit is the BaseFitter template class that provides custom fitting to any definable function with minimal boilerplate code. All the plotting and boundary functionalities are handled inside the template class so that the user can focus solely on running through multiple models quickly without thinking about how to manage multiple models of the same type or even of different types.

Additionally, the generators template function provides the user with an N-model data generator function with added noise capability to mimic real-life scenarios of whatever distribution the user might want.

      - #221
      - @SimonMolinsky
