### 2025-01-29 #233 @coatless
---
---

I think there's enough novelty behind the multifitter and [distribution](https://pymultifit.readthedocs.io/latest/basic_examples/basics1.html) approaches discussed to move forward with a full review. (For clarity, `scipy` provides [`scipy.stats.fit()`](https://docs.scipy.org/doc/scipy-1.15.1/reference/generated/scipy.stats.fit.html) for a single DV or CV whereas multiple and different supports are given by `pymultifit` via [`MixedDataFitter`](https://pymultifit.readthedocs.io/latest/basic_examples/basics4.html)) Moreover, there are ample tutorials and a solid case study of applying the package to solve real-world problems.

My only concern is regarding the accuracy benchmarks throwing `RuntimeWarning` notices. I think this is leaning into a discussion that appeared in the presubmission intake analysis regarding a re-implementation with `numpy`.

For example, with `arcsine()` and `beta(a=5, b=80, loc=-3, scale=5)` we have:

### Beta

https://pymultifit.readthedocs.io/latest/benchmarks/_bm_accuracy.html#beta(a=5,-b=80,-loc=-3,-scale=5)

```py
/home/sarl-ws-5/PycharmProjects/pyMultiFit/src/pymultifit/distributions/utilities_d.py:168: RuntimeWarning: overflow encountered in power
  numerator = y**(alpha - 1) * (1 - y)**(beta - 1)
/home/sarl-ws-5/PycharmProjects/pyMultiFit/src/pymultifit/distributions/utilities_d.py:168: RuntimeWarning: overflow encountered in multiply
  numerator = y**(alpha - 1) * (1 - y)**(beta - 1)
```

### Arcsine

https://pymultifit.readthedocs.io/latest/benchmarks/_bm_accuracy.html#arcsine()

```py
/home/sarl-ws-5/PycharmProjects/pyMultiFit/src/pymultifit/distributions/utilities_d.py:71: RuntimeWarning: invalid value encountered in sqrt
  pdf_ = 1 / (np.pi * np.sqrt(y * (1 - y)))
/home/sarl-ws-5/PycharmProjects/pyMultiFit/src/pymultifit/distributions/utilities_d.py:71: RuntimeWarning: divide by zero encountered in divide
  pdf_ = 1 / (np.pi * np.sqrt(y * (1 - y)))
/home/sarl-ws-5/PycharmProjects/pyMultiFit/benchmarks/functions.py:30: RuntimeWarning: invalid value encountered in subtract
  pdf_abs_diff = np.abs(pdf_custom - pdf_scipy) + EPSILON
```
