### 2025-01-29 517e6fe
[main-minor-10]
1. benchmarks
    1. updated `speed.ipynb` and `accuracy.ipynb`
2. src/pymultifit/distributions
    1. updated `utilities_d.py` for `arc_sine_pdf_` and `beta_pdf_` to avoid run-time errors and power overflow errors

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
