+++ src/tests/test_distributions/test_arcSine_d.py
@@ -20,8 +20,10 @@ class TestArcSineDistribution:
         x = np.linspace(start=0, stop=1, num=100)
         _distribution1 = ArcSineDistribution(normalize=True)
         _distribution2 = beta.pdf(x, a=0.5, b=0.5)
+        _distribution3 = arcsine.pdf(x)
 
         np.testing.assert_allclose(actual=_distribution1.pdf(x), desired=_distribution2, rtol=1e-5, atol=1e-8)
+        np.testing.assert_allclose(actual=_distribution1.pdf(x), desired=_distribution3, rtol=1e-5, atol=1e-8)
 
     @staticmethod
     def test_stats():
@@ -46,7 +48,7 @@ class TestArcSineDistribution:
 
     @staticmethod
     def test_pdf_cdf():
-        x1 = np.linspace(start=-0.5, stop=1.5, num=10)
+        x1 = np.linspace(start=-0.5, stop=1.5, num=100)
         x2 = np.array([0, 1])
 
         for x in [x1, x2]:
@@ -60,8 +62,8 @@ class TestArcSineDistribution:
             np.testing.assert_allclose(actual=pdf_custom, desired=pdf_scipy, rtol=1e-5, atol=1e-8)
             np.testing.assert_allclose(actual=cdf_custom, desired=cdf_scipy, rtol=1e-5, atol=1e-8)
 
-        loc_ = np.random.uniform(low=-5, high=10, size=50)
-        scale_ = np.random.uniform(low=EPSILON, high=10, size=50)
+        loc_ = np.random.uniform(low=-5, high=10, size=100)
+        scale_ = np.random.uniform(low=EPSILON, high=10, size=100)
         stack_ = np.column_stack([loc_, scale_])
 
         for loc, scale in stack_:
