### 2025-01-29 #233 @syedalimohsinbukhari
Hi @coatless

Thanks for getting back. I've been meaning to address those issues for a while but was waiting for a response before I proceeded. I've updated the required functions, and they shouldn't give the same issues now. The only issue now is that in `arcsine` for x=1 (in the edge case of testing), the pdf gives `np.inf` in both scenarios and thus the invalid value error.

![Image](https://github.com/user-attachments/assets/01a9c47f-fb44-4116-b131-aacd52113bb1)

Cheers.
