### 2025-01-29 #233 @coatless
@syedalimohsinbukhari any reason for not directly using [`np.arcsin()`](https://numpy.org/doc/stable/reference/generated/numpy.arcsin.html)? I'm not having that issue with the built-in version:

---

Could you restore the docstring for `beta_pdf_()`:

https://github.com/syedalimohsinbukhari/pyMultiFit/commit/517e6fe30106216dbbd7d7c9e1bd1a2ce4f587b4#diff-901f230daf84fc7cccf4819b150e039e69d3b060a1bcc2279a55f84ce2570da6L118
