### 2025-01-29 9323e2f
[main-minor-11]
1. src/pymultifit/distributions
    1. updated `utilities_d.py` with beta_pdf_ docstring
    2. added `genNorm_d.py` (still testing)

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
