+++ src/pymultifit/distributions/generalized/__init__.py
@@ -1,0 +1,1 @@
+"""Created on Jan 29 15:42:06 2025"""
