### 2025-01-29 #233 @syedalimohsinbukhari
Hi @coatless,

I think this image should clear things up. The `ArcSineDistribution` is not the same as `np.arcsin` of trigonometry; it is a distribution with pdf $f(y) = \dfrac{1}{\pi\sqrt{y(1-y)}}$ that's why.

For reference, [wiki article](https://en.wikipedia.org/wiki/Arcsine_distribution) and [scipy docs](https://docs.scipy.org/doc/scipy/tutorial/stats/continuous_arcsine.html).

![Image](https://github.com/user-attachments/assets/fa632b2a-c16c-475a-b763-f74b54c898ae)

Thanks for the headsup for `beta_pdf_`; strangely, it is showing up in my local builds but not in recent RTD builds. I'll look into it more.

![Image](https://github.com/user-attachments/assets/5ee73de8-9d7a-4cad-8d48-94aa2bfd7c83)
