### 2025-01-30 #233 @syedalimohsinbukhari
Hi @coatless

Just letting you know that `beta_pdf_()` docstring is now showing up on RTD servers as well.

https://pymultifit.readthedocs.io/latest/distributions/utilities_d.html#pymultifit.distributions.utilities_d.beta_pdf_
