### 2025-01-31 8948ace
[gen-norm-minor-1]
1. added `genNorm_d.py`
2. added variance in  `foldedNormal_d.py`
3. added `gen_sym_normal_pdf_`, `gen_sym_normal_cdf_`, `un_norm_gamma_inc_low` and `un_norm_gamma_inc_upp` to `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
