+++ src/pymultifit/distributions/generalized/genNorm_d.py
@@ -1,12 +1,12 @@
 """Created on Jan 29 15:42:23 2025"""
 
-import numpy as np
-from scipy.special.cython_special import gammaincc, gammainc
-from scipy.stats import gennorm
+from typing import Dict
 
-from pymultifit import EPSILON
-from pymultifit.distributions.backend import BaseDistribution
-from pymultifit.distributions.utilities_d import _pdf_scaling
+import numpy as np
+from scipy.special import gamma
+
+from ..backend import BaseDistribution
+from ..utilities_d import gen_sym_normal_pdf_, gen_sym_normal_cdf_
 
 
 class SymmetricGeneralizedGaussianDistribution(BaseDistribution):
@@ -15,7 +15,7 @@ class SymmetricGeneralizedGaussianDistribution(BaseDistribution):
                  normalize: bool = False):
         self.amplitude = 1. if normalize else amplitude
         self.mu = loc
-        self.alpha1 = scale
+        self.alpha = scale
         self.beta = shape
 
         self.norm = normalize
@@ -26,84 +26,22 @@ class SymmetricGeneralizedGaussianDistribution(BaseDistribution):
         return instance
 
     def pdf(self, x: np.ndarray) -> np.ndarray:
-        return gen_gaussian_pdf_(x,
-                                 amplitude=self.amplitude, loc=self.mu, scale=self.alpha1, shape=self.beta,
-                                 skew=self.alpha1, normalize=self.norm)
+        return gen_sym_normal_pdf_(x, amplitude=self.amplitude, loc=self.mu, scale=self.alpha, shape=self.beta,
+                                   normalize=self.norm)
 
     def cdf(self, x: np.ndarray) -> np.ndarray:
-        return gen_gaussian_cdf_(x,
-                                 amplitude=self.amplitude, loc=self.mu, scale=self.alpha1, shape=self.beta,
-                                 skew=self.alpha1, normalize=self.norm)
+        return gen_sym_normal_cdf_(x, amplitude=self.amplitude, loc=self.mu, scale=self.alpha, shape=self.beta,
+                                   normalize=self.norm)
 
+    def stats(self) -> Dict[str, float]:
+        mean_ = self.mu
+        median_ = self.mu
+        mode_ = self.mu
+        variance_ = self.alpha**2 * gamma(3 / self.beta)
+        variance_ /= gamma(1 / self.beta)
 
-def gen_gaussian_pdf_(x: np.ndarray,
-                      amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, shape: float = 1.0,
-                      skew: float = 1.0, normalize: bool = False) -> np.ndarray:
-    # taken from https://ieeexplore.ieee.org/document/8959693
-    mu, alpha1, beta, alpha2 = loc, scale, shape, skew
-
-    f1 = (-x + mu) / alpha1
-    f2 = (x - mu) / alpha2
-    denominator = (alpha1 + alpha2) * gamma(1 / beta)
-    numerator = np.where(x < mu, np.exp(-f1**beta), np.exp(-f2**beta))
-    pdf_ = (beta * numerator) / denominator
-
-    if not normalize:
-        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
-
-    return pdf_
-
-
-import numpy as np
-from scipy.special import gamma, gammaincc, gammainc
-
-
-def gen_gaussian_cdf_(x: np.ndarray, amplitude: float = 1.0, loc: float = 0.0,
[…]
