### 2025-01-31 7022a34
[gen-norm-minor-2]
1. src/pymultifit/distributions
    1. updated `genNorm_d.py`
    2. added additional functionalities in `utilities_d.py` for genNorm
2. tests
    1. added `test_gen_norm.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
