### 2025-02-01 b25fe1a
[gen-norm-major-1]
1. removed a lot of numpy overhead from `distributions` module
2. retested for speed and accuracy
3. added `gennorm.py` in examples
4.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
