+++ src/pymultifit/distributions/arcSine_d.py
@@ -1,11 +1,7 @@
 """Created on Aug 14 02:02:42 2024"""
 
-from typing import Dict
-
-import numpy as np
-
 from .backend import errorHandling as erH, BaseDistribution
-from .utilities_d import arc_sine_pdf_, arc_sine_cdf_
+from .utilities_d import arc_sine_pdf_, arc_sine_cdf_, arc_sine_log_pdf_, arc_sine_log_cdf_
 
 
 class ArcSineDistribution(BaseDistribution):
@@ -60,7 +56,7 @@ class ArcSineDistribution(BaseDistribution):
        :language: python
        :linenos:
        :lineno-start: 14
-       :lines: 14-
+       :lines: 14-27
 
     .. image:: ../../../images/arcsine_example.png
        :alt: ArcSine distribution
@@ -98,13 +94,19 @@ class ArcSineDistribution(BaseDistribution):
         """
         return cls(loc=loc, scale=scale, normalize=True)
 
-    def pdf(self, x: np.ndarray) -> np.ndarray:
+    def pdf(self, x):
         return arc_sine_pdf_(x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
 
-    def cdf(self, x: np.ndarray) -> np.ndarray:
+    def logpdf(self, x):
+        return arc_sine_log_pdf_(x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
+
+    def cdf(self, x):
         return arc_sine_cdf_(x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
 
-    def stats(self) -> Dict[str, float]:
+    def logcdf(self, x):
+        return arc_sine_log_cdf_(x, amplitude=self.amplitude, loc=self.loc, scale=self.scale, normalize=self.norm)
+
+    def stats(self):
         s_, l_ = self.scale, self.loc
 
         mean_ = (s_ * 0.5) + l_
@@ -115,4 +117,4 @@ class ArcSineDistribution(BaseDistribution):
                 'median': median_,
                 'mode': None,
                 'variance': variance_,
-                'std': np.sqrt(variance_)}
+                'std': variance_**0.5}
