### 2025-02-01 87f8d6b
[gen-norm-minor-3]
1. minor changes in `genNorm_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
