### 2025-02-02 9116125
[log-distributions-minor-1]
1. added `scaledInvChiSquare_d.py`
2. added test for `test_scaledInvChiSquare_d.py`
3. updated `__init__.py`, `_distributions.rst`, `scaledInvChiSquare_d.rst`, and `utilities_d.rst`
4. added `pdf`, `cdf`, `log_pdf`, and `log_cdf` functionality in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
