+++ src/tests/test_distributions/test_scaledInvChiSquare_d.py
@@ -1,0 +1,138 @@
+"""Created on Feb 02 17:13:40 2025"""
+
+import numpy as np
+import pytest
+from scipy.stats import invgamma
+
+from ...pymultifit import EPSILON
+from ...pymultifit.distributions import ScaledInverseChiSquareDistribution
+from ...pymultifit.distributions.backend import errorHandling as erH
+
+
+class TestGaussianDistribution:
+
+    @staticmethod
+    def test_initialization():
+        dist = ScaledInverseChiSquareDistribution(amplitude=2.0, df=1.0, scale=0.5, normalize=False)
+        assert dist.amplitude == 2.0
+        assert dist.df == 1.0
+        assert dist.scale == 0.5
+        assert dist.tau2 == dist.scale / dist.df
+        assert not dist.norm
+
+        dist_normalized = ScaledInverseChiSquareDistribution(amplitude=2.0, normalize=True)
+        assert dist_normalized.amplitude == 1.0
+
+    @staticmethod
+    def test_constraints():
+        with pytest.raises(erH.NegativeAmplitudeError, match=f"Amplitude {erH.neg_message}"):
+            ScaledInverseChiSquareDistribution(amplitude=-1.0, normalize=False)
+
+        # amplitude should be internally updated to 1.0 if `normalize` is called
+        distribution = ScaledInverseChiSquareDistribution(amplitude=-1.0, normalize=True)
+        assert distribution.amplitude == 1.0
+
+        with pytest.raises(erH.NegativeScaleError, match=f"Scale {erH.neg_message}"):
+            ScaledInverseChiSquareDistribution(scale=-3.0)
+
+    @staticmethod
+    def test_edge_case():
+        dist = ScaledInverseChiSquareDistribution()
+        x = np.array([])
+        result = dist.pdf(x)
+        assert result.size == 0
+
+        result = dist.cdf(x)
+        assert result.size == 0
+
+        result = dist.logpdf(x)
+        assert result.size == 0
+
+        result = dist.logcdf(x)
+        assert result.size == 0
+
+    @staticmethod
+    def test_stats():
+        beta_ = np.random.uniform(low=EPSILON, high=10, size=10)
+        loc_ = np.random.uniform(low=-5, high=10, size=10)
+        scale_ = np.random.uniform(low=EPSILON, high=10, size=10)
+        stack_ = np.column_stack([beta_, loc_, scale_])
+
+        for beta, loc, scale in stack_:
+            _distribution = ScaledInverseChiSquareDistribution.scipy_like(a=beta, loc=loc, scale=scale)
+            d_stats = _distribution.stats()
+
+            # Scipy calculations
+            scipy_mean, scipy_variance = invgamma.stats(a=beta / 2, loc=loc, scale=scale / 2, moments='mv')
+            scipy_stddev = np.sqrt(scipy_variance)
+
+            # Assertions for mean and variance
+            np.testing.assert_allclose(actual=scipy_mean, desired=d_stats['mean'], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_variance, desired=d_stats['variance'], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_stddev, desired=d_stats['std'], rtol=1e-5, atol=1e-8)
+
+    @staticmethod
+    def test_pdf_cdf():
+        df = np.random.uniform(low=1, high=100, size=1000)
[…]
