+++ src/pymultifit/distributions/utilities_d.py
@@ -13,6 +13,8 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
            'half_normal_pdf_', 'half_normal_cdf_',
            'laplace_pdf_', 'laplace_cdf_',
            'log_normal_pdf_', 'log_normal_cdf_',
+           'scaled_inv_chi_square_pdf_', 'scaled_inv_chi_square_log_pdf_',
+           'scaled_inv_chi_square_cdf_', 'scaled_inv_chi_square_log_cdf_',
            'skew_normal_pdf_', 'skew_normal_cdf_',
            'uniform_pdf_', 'uniform_cdf_']
 
@@ -20,7 +22,7 @@ from typing import Union
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t
+from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t, gammaincc
 
 from .. import doc_style
 
@@ -1162,6 +1164,149 @@ def uniform_cdf_(x: np.ndarray,
     return cdf_values
 
 
+def scaled_inv_chi_square_pdf_(x, amplitude: float = 1.0, df: float = 1.0, scale: float = 1.0,
+                               loc: float = 0.0, normalize: bool = False):
+    r"""
+    Compute PDF of :class:`~pymultifit.distributions.scaledInvChiSquare_d.ScaledInverseChiSquareDistribution`.
+
+    Parameters
+    ----------
+    x : np.ndarray
+        Input array of values.
+    amplitude : float, optional
+        The amplitude of the PDF.
+        Defaults to 1.0.
+        Ignored if **normalize** is ``True``.
+    df : float, optional
+        The degree of freedom.
+        Defaults to 1.0.
+    scale: float, optional
+        The scale parameter, for scaling.
+        Defaults to 1.0,
+    loc : float, optional
+        The location parameter, for shifting.
+        Defaults to 0.0.
+    normalize : bool, optional
+        If ``True``, the distribution is normalized so that the total area under the PDF equals 1.
+        Defaults to ``False``.
+
+    Returns
+    -------
+    np.ndarray
+        Array of the same shape as :math:`x`, containing the evaluated values.
+
+    Notes
+    -----
+    The Scaled Inverse ChiSquare PDF is defined as:
+
+    .. math:: f(y\ | \nu,\phi) = \dfrac{\tau^2\nu_2}{\Gamma(\nu_2)}\dfrac{1}{y^{1+\nu_2}}\exp\left[-\dfrac{\nu\tau^2}{2y}\right]
+
+    where :math:`\nu_2 = \dfrac{\nu}{2}`, :math:`\tau^2 = \dfrac{\phi}{\nu}` and :math:`y` is the transformed
+    value of :math:`x`, defined as:
+
+    .. math:: y = x - \text{loc}
+
+    The final PDF is expressed as :math:`f(y)`.
+    """
+    if x.size == 0:
+        return np.array([])
+
+    tau2 = scale / df
+    df_half = df / 2
+
+    y = x - loc
+
+    pdf_ = np.zeros_like(a=y, dtype=float)
+    mask = y > 0
+
+    frac1 = np.power(tau2 * df_half, df_half) / gamma(df_half)
+    if np.any(mask):
+        y_valid = y[mask]
+        frac2 = np.exp(-(df * tau2) / (2 * y_valid)) / np.power(y_valid, 1 + df_half)
+
[…]

[…]
