### 2025-02-02 306dc84
[log-distributions-minor-1-2]
1. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
