+++ docs/source/distributions/arcSine_d.rst
@@ -7,7 +7,7 @@ ArcSine Distribution
    :members:
    :inherited-members:
    :show-inheritance:
-   :member-order: groupwise
+   :member-order: bysource
 
 
 This class internally utilizes the following functions from :mod:`~pymultifit.distributions.utilities_d` module:
