### 2025-02-02 8be14de
[log-distributions-minor-2]
1. added `base_test_functions.py` for easy testing utilities
2. updated `test_arcSine_d.py` and `test_beta_d.py`
3. updated `utilities_d.py` and `utilities_d.rst`
4. added `statsmodels` to `environment.yaml` and `requirements[dev].txt`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
