+++ src/tests/test_distributions/base_test_functions.py
@@ -1,0 +1,98 @@
+"""Created on Feb 02 22:24:06 2025"""
+
+import numpy as np
+
+from pymultifit import EPSILON
+
+loc_parameter = np.random.uniform(low=-500, high=500, size=1000)
+scale_parameter = np.random.uniform(low=EPSILON, high=500, size=1000)
+shape_parameter = np.random.uniform(low=EPSILON, high=500, size=1000)
+
+
+def edge_cases(distribution, log_check=False):
+    x = np.array([])
+
+    result = distribution.pdf(x)
+    assert result.size == 0
+    result = distribution.cdf(x)
+    assert result.size == 0
+
+    if log_check:
+        result = distribution.logpdf(x)
+        assert result.size == 0
+        result = distribution.logcdf(x)
+        assert result.size == 0
+
+
+def scaled_distributions(custom_distribution, scipy_distribution, x, parameters, log_check: bool = False):
+    dist = custom_distribution(*parameters)
+
+    np.testing.assert_allclose(actual=dist.pdf(x), desired=scipy_distribution(*parameters).pdf(x), rtol=1e-5, atol=1e-8)
+    np.testing.assert_allclose(actual=dist.cdf(x), desired=scipy_distribution(*parameters).cdf(x), rtol=1e-5, atol=1e-8)
+    if log_check:
+        np.testing.assert_allclose(actual=dist.logpdf(x), desired=scipy_distribution(*parameters).logpdf(x), rtol=1e-5,
+                                   atol=1e-8)
+        np.testing.assert_allclose(actual=dist.logcdf(x), desired=scipy_distribution(*parameters).logcdf(x), rtol=1e-5,
+                                   atol=1e-8)
+
+
+def statistics(custom_distribution, scipy_distribution, parameters, mean_variance=False, median=False):
+    d_stats = custom_distribution(*parameters).stats()
+
+    if mean_variance:
+        scipy_mean, scipy_variance = scipy_distribution.stats(*parameters, moments='mv')
+        scipy_stddev = np.sqrt(scipy_variance)
+        np.testing.assert_allclose(actual=scipy_mean, desired=d_stats['mean'], rtol=1e-5, atol=1e-8)
+        np.testing.assert_allclose(actual=scipy_variance, desired=d_stats['variance'], rtol=1e-5, atol=1e-8)
+        np.testing.assert_allclose(actual=scipy_stddev, desired=d_stats['std'], rtol=1e-5, atol=1e-8)
+
+    if median:
+        scipy_median = scipy_distribution.median(*parameters)
+        np.testing.assert_allclose(actual=scipy_median, desired=d_stats['median'], rtol=1e-5, atol=1e-8)
+
+
+def stats(custom_distribution, scipy_distribution, parameters, mean_variance=True, median=True):
+    stack_ = np.column_stack(parameters)
+
+    for stack in stack_:
+        statistics(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
+                   parameters=stack, mean_variance=mean_variance, median=median)
+
+
+def value_functions(custom_distribution, scipy_distribution, parameters, n_values=10):
+    x = np.linspace(start=-100, stop=100, num=n_values)
+    x = np.concatenate([x, np.array([0, 1])])
+    x.sort()
+
+    stack_ = np.column_stack(parameters)
+
+    for params in stack_:
+        scaled_distributions(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
+                             x=x, parameters=params)
+
+
[…]
