+++ docs/source/distributions/utilities_d.rst
@@ -12,7 +12,9 @@ Available for user
 ------------------
 
 .. autofunction:: arc_sine_pdf_
+.. autofunction:: arc_sine_log_pdf_
 .. autofunction:: arc_sine_cdf_
+.. autofunction:: arc_sine_log_cdf_
 .. autofunction:: beta_pdf_
 .. autofunction:: beta_cdf_
 .. autofunction:: chi_square_pdf_
