### 2025-02-03 ab805a1
[log-distributions-minor-3]
1. correction in `base_test_functions.py`
2. updated `beta_d.py` with `logpdf` and `logcdf`
3. updated `test_beta_d.py`, `utilities_d.rst`
4. added beta logpdf and logcdf in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
