+++ src/pymultifit/distributions/beta_d.py
@@ -3,7 +3,7 @@
 from scipy.special import betaincinv
 
 from .backend import BaseDistribution, errorHandling as erH
-from .utilities_d import beta_cdf_, beta_pdf_
+from .utilities_d import beta_cdf_, beta_pdf_, beta_log_pdf_, beta_log_cdf_
 
 
 class BetaDistribution(BaseDistribution):
@@ -122,11 +122,19 @@ class BetaDistribution(BaseDistribution):
         return cls(alpha=a, beta=b, loc=loc, scale=scale, normalize=True)
 
     def pdf(self, x):
-        return beta_pdf_(x=x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, loc=self.loc,
+        return beta_pdf_(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, loc=self.loc,
                          scale=self.scale, normalize=self.norm)
 
+    def logpdf(self, x):
+        return beta_log_pdf_(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, loc=self.loc,
+                             scale=self.scale, normalize=self.norm)
+
     def cdf(self, x):
-        return beta_cdf_(x=x, alpha=self.alpha, beta=self.beta, loc=self.loc, scale=self.scale)
+        return beta_cdf_(x, alpha=self.alpha, beta=self.beta, loc=self.loc, scale=self.scale)
+
+    def logcdf(self, x):
+        return beta_log_cdf_(x, amplitude=self.amplitude, alpha=self.alpha, beta=self.beta, loc=self.loc,
+                             scale=self.scale, normalize=self.norm)
 
     def stats(self):
         a, b = self.alpha, self.beta
