### 2025-02-03 59538be
[log-distributions-minor-4]
1. added more testing to `base_test_functions.py`
2. minor corrections in `baseDistribution.py`
3. updated `test_arcSine_d.py`, `test_beta_d.py` and `test_chiSquare_d.py`
4. chi2 pdf and cdf corrected to logpdf logcdf in `utilities_d.py`
