+++ src/pymultifit/distributions/chiSquare_d.py
@@ -3,7 +3,7 @@
 from typing import Union
 
 from .backend import BaseDistribution, errorHandling as erH
-from .utilities_d import chi_square_cdf_, chi_square_pdf_
+from .utilities_d import chi_square_cdf_, chi_square_pdf_, chi_square_log_pdf_, chi_square_log_cdf_
 
 
 class ChiSquareDistribution(BaseDistribution):
@@ -116,9 +116,17 @@ class ChiSquareDistribution(BaseDistribution):
         return chi_square_pdf_(x, amplitude=self.amplitude, degree_of_freedom=self.dof, loc=self.loc, scale=self.scale,
                                normalize=self.norm)
 
+    def logpdf(self, x):
+        return chi_square_log_pdf_(x, amplitude=self.amplitude, degree_of_freedom=self.dof, loc=self.loc,
+                                   scale=self.scale, normalize=self.norm)
+
     def cdf(self, x):
         return chi_square_cdf_(x, amplitude=self.amplitude, degree_of_freedom=self.dof, loc=self.loc, scale=self.scale,
                                normalize=self.norm)
+
+    def logcdf(self, x):
+        return chi_square_log_cdf_(x, amplitude=self.amplitude, degree_of_freedom=self.dof, loc=self.loc,
+                                   scale=self.scale, normalize=self.norm)
 
     def stats(self):
         df = self.dof
