+++ src/pymultifit/distributions/utilities_d.py
@@ -116,7 +116,7 @@ def arc_sine_log_pdf_(x: np.ndarray,
     log_pdf_[y == 0] = np.inf
     log_pdf_[y == 1] = np.inf
 
-    return _remove_nans(log_pdf_ - np.log(scale))
+    return log_pdf_ - np.log(scale)
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
@@ -186,7 +186,7 @@ def arc_sine_log_cdf_(x: np.array,
     log_cdf_[mask_] = np.log(2 / np.pi) + np.log(np.arcsin(np.sqrt(y[mask_])))
     log_cdf_[y >= 1] = np.log(1)
 
-    return _remove_nans(log_cdf_)
+    return log_cdf_
 
 
 def beta_pdf_(x: np.ndarray,
@@ -236,6 +236,9 @@ def beta_pdf_(x: np.ndarray,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
@@ -262,13 +265,14 @@ def beta_pdf_(x: np.ndarray,
 
     # handle the cases where nans can occur with nan_to_num
     # np.inf and -np.inf to not affect the infinite values
-    return _remove_nans(pdf_ / scale)
+    pdf_ = _remove_nans(pdf_ / scale)
+    return pdf_.item() if scalar_input else pdf_
 
 
 @doc_inherit(parent=beta_pdf_, style=doc_style)
 def beta_log_pdf_(x: np.ndarray,
-                 amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
-                 loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
+                  amplitude: float = 1.0, alpha: float = 1.0, beta: float = 1.0,
+                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
     r"""Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Notes
@@ -284,33 +288,38 @@ def beta_log_pdf_(x: np.ndarray,
 
     The final logPDF is expressed as :math:`\ell(y) - \ln(\text{scale})`.
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
     y = (x - loc) / scale
 
-    logpdf_ = np.full_like(a=y, fill_value=-np.inf, dtype=float)
+    log_pdf_ = np.full_like(a=y, fill_value=-np.inf, dtype=float)
     mask_ = ~_beta_masking(y=y, alpha=alpha, beta=beta)
 
     log_numerator = np.zeros_like(a=y, dtype=float)
     log_numerator[mask_] = (alpha - 1) * np.log(y[mask_]) + (beta - 1) * np.log(1 - y[mask_])
     normalization_factor = gammaln(alpha) + gammaln(beta) - gammaln(alpha + beta)
 
-    logpdf_[mask_] = log_numerator[mask_] - normalization_factor
+    log_pdf_[mask_] = log_numerator[mask_] - normalization_factor
 
     if alpha <= 1:
-        logpdf_[y == 0] = np.nan
+        log_pdf_[y == 0] = np.nan
     if beta <= 1:
-        logpdf_[y == 1] = np.nan
+        log_pdf_[y == 1] = np.nan
     if alpha == 1 and beta == 1:
-        logpdf_[y == 1] = np.log(1)
[…]
