### 2025-02-03 8dcdd91
[log-distributions-minor-4-1]
1. minor corrections in tests
