### 2025-02-03 6ba1f5b
[log-distributions-minor-4-2]
1. minor corrections in `baseDistribution.py`
2. minor changes
