### 2025-02-04 75454cf
[log-distributions-minor-5]
1. modified `base_test_functions.py` for exponential distribution
2. added logpdf and logcdf to `exponential_d.py`
3. updated `test_exponential_d.py`
4. minor changes in `utilities_d.py`
