+++ src/tests/test_distributions/base_test_functions.py
@@ -24,8 +24,13 @@ def edge_cases(distribution, log_check=False):
         assert result.size == 0
 
 
-def scaled_distributions(custom_distribution, scipy_distribution, x, parameters, log_check: bool = False):
+def scaled_distributions(custom_distribution, scipy_distribution, x, parameters, log_check: bool = False,
+                         is_expon=False):
     dist = custom_distribution(*parameters)
+
+    # make an exception for exponential distribution which gets scale = 1/scale
+    if is_expon:
+        parameters[1] = 1 / parameters[1]
 
     np.testing.assert_allclose(actual=dist.pdf(x), desired=scipy_distribution(*parameters).pdf(x), rtol=1e-5, atol=1e-8)
     np.testing.assert_allclose(actual=dist.cdf(x), desired=scipy_distribution(*parameters).cdf(x), rtol=1e-5, atol=1e-8)
@@ -36,8 +41,12 @@ def scaled_distributions(custom_distribution, scipy_distribution, x, parameters,
                                    atol=1e-8)
 
 
-def statistics(custom_distribution, scipy_distribution, parameters, mean_variance=False, median=False):
+def statistics(custom_distribution, scipy_distribution, parameters, mean_variance=False, median=False, is_expon=False):
     d_stats = custom_distribution(*parameters).stats()
+
+    # make an exception for exponential distribution which gets scale = 1/scale
+    if is_expon:
+        parameters[1] = 1 / parameters[1]
 
     if mean_variance:
         scipy_mean, scipy_variance = scipy_distribution.stats(*parameters, moments='mv')
@@ -51,15 +60,15 @@ def statistics(custom_distribution, scipy_distribution, parameters, mean_varianc
         np.testing.assert_allclose(actual=scipy_median, desired=d_stats['median'], rtol=1e-5, atol=1e-8)
 
 
-def stats(custom_distribution, scipy_distribution, parameters, mean_variance=True, median=True):
+def stats(custom_distribution, scipy_distribution, parameters, mean_variance=True, median=True, is_expon=False):
     stack_ = np.column_stack(parameters)
 
     for stack in stack_:
         statistics(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
-                   parameters=stack, mean_variance=mean_variance, median=median)
+                   parameters=stack, mean_variance=mean_variance, median=median, is_expon=is_expon)
 
 
-def value_functions(custom_distribution, scipy_distribution, parameters, n_values=10, log_check=False):
+def value_functions(custom_distribution, scipy_distribution, parameters, n_values=10, log_check=False, is_expon=False):
     x = np.linspace(start=-100, stop=100, num=n_values)
     x = np.concatenate([x, np.array([0, 1])])
     x.sort()
@@ -68,7 +77,7 @@ def value_functions(custom_distribution, scipy_distribution, parameters, n_value
 
     for params in stack_:
         scaled_distributions(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
-                             x=x, parameters=params, log_check=log_check)
+                             x=x, parameters=params, log_check=log_check, is_expon=is_expon)
 
 
 def two_parameter_extreme_test(custom_distribution, scipy_distribution, log_check=False):
@@ -98,15 +107,20 @@ def two_parameter_extreme_test(custom_distribution, scipy_distribution, log_chec
             np.testing.assert_allclose(actual=dist_.logcdf(x_), desired=scipy_logcdf, rtol=1e-5, atol=1e-8)
 
 
-def single_input_n_variables(custom_distribution, scipy_distribution, parameters, n_size=1_000, log_check=False):
[…]
