### 2025-02-04 ca6f3ad
[log-distributions-minor-5-1]
1. minor correction in `base_test_functions.py`
