### 2025-02-04 5dbb11d
[log-distributions-minor-6]
1. added logpdf to `foldedNormal_d.py`
2. minor changes in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
