### 2025-02-04 876a8a2
[log-distributions-minor-7]
1. added logcdf to `foldedNormal_d.py` and updated stats
2. updated `test_foldedNormal_d.py`
3. updated folded_normal_log_pdf_ calculations in `utilities_d.py` and added `folded_normal_log_cdf_`
