### 2025-02-05 2da2927
[log-distributions-minor-7]
1. minor updates in `arcSine_d.py`, `beta_d.py`, `chiSquare_d.py`, `exponential_d.py`, and `foldedNormal_d.py`
2. updated `gamma_d.py` with logcdf and logpdf as well
3. updated `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
