### 2025-02-06 2ed7b22
[log-distributions-minor-8]
1. added logpdf and logcdf to `gaussian_d.py`
2. updated `test_gaussian_d.py` and `utilities_d.py`
