+++ src/tests/test_distributions/test_gaussian_d.py
@@ -4,9 +4,11 @@ import numpy as np
 import pytest
 from scipy.stats import norm
 
-from ...pymultifit import EPSILON
+from . import base_test_functions as btf
 from ...pymultifit.distributions import GaussianDistribution
 from ...pymultifit.distributions.backend import errorHandling as erH
+
+np.random.seed(45)
 
 
 class TestGaussianDistribution:
@@ -35,85 +37,22 @@ class TestGaussianDistribution:
             GaussianDistribution(std=-3.0)
 
     @staticmethod
-    def test_edge_case():
-        dist = GaussianDistribution()
-        x = np.array([])
-        result = dist.pdf(x)
-        assert result.size == 0  # Should return an empty array
+    def test_edge_cases():
+        btf.edge_cases(distribution=GaussianDistribution(), log_check=True)
 
     @staticmethod
     def test_stats():
-        loc_ = np.random.uniform(low=-5, high=10, size=10)
-        scale_ = np.random.uniform(low=EPSILON, high=10, size=10)
-        stack_ = np.column_stack([loc_, scale_])
-
-        for loc, scale in stack_:
-            _distribution = GaussianDistribution.scipy_like(loc=loc, scale=scale)
-            d_stats = _distribution.stats()
-
-            # Scipy calculations
-            scipy_mean, scipy_variance = norm.stats(loc=loc, scale=scale, moments='mv')
-            scipy_median = norm.median(loc=loc, scale=scale)
-            scipy_stddev = np.sqrt(scipy_variance)
-
-            # Assertions for mean and variance
-            np.testing.assert_allclose(actual=scipy_mean, desired=d_stats['mean'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_variance, desired=d_stats['variance'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_median, desired=d_stats['median'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_stddev, desired=d_stats['std'], rtol=1e-5, atol=1e-8)
+        btf.stats(custom_distribution=GaussianDistribution.scipy_like, scipy_distribution=norm,
+                  parameters=[btf.loc_parameter, btf.scale_parameter])
 
     @staticmethod
-    def test_pdf_cdf():
-        def _cdf_pdf_custom(x_, dist_, what='cdf'):
-            return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
-
-        def _cdf_pdf_scipy(x_, loc, scale, what='cdf'):
-            return norm.cdf(x_, loc=loc, scale=scale) if what == 'cdf' else norm.pdf(x_, loc=loc, scale=scale)
-
-        for i in ['cdf', 'pdf']:
-            for _ in range(50):  # Run 50 random tests
-                loc_ = np.random.uniform(low=-10, high=10)
-                scale_ = np.random.uniform(low=0.1, high=5)
-                x = np.linspace(start=loc_ - 10, stop=loc_ + 10, num=50)
-                distribution = GaussianDistribution(mu=loc_, std=scale_, normalize=True)
-                expected = _cdf_pdf_scipy(x_=x, loc=loc_, scale=scale_, what=i)
-                np.testing.assert_allclose(actual=_cdf_pdf_custom(x_=x, dist_=distribution, what=i),
-                                           desired=expected, rtol=1e-5, atol=1e-8)
+    def test_pdfs():
+        btf.value_functions(custom_distribution=GaussianDistribution.scipy_like, scipy_distribution=norm,
+                            parameters=[btf.loc_parameter, btf.scale_parameter], log_check=True)
 
     @staticmethod
[…]
