### 2025-02-06 5239071
[log-distributions-minor-9]
1. updated `halfNormal_d.py`, `laplace_d.py` and `logNormal_d.py`
