### 2025-02-07 0be2684
[log-distributions-minor-9-1]
1. more `logNormal_d.py` testing
