+++ src/tests/test_distributions/test_logNormal_d.py
@@ -45,16 +45,16 @@ class TestLogNormalDistribution:
     @staticmethod
     def test_stats():
         btf.stats(custom_distribution=LogNormalDistribution.scipy_like, scipy_distribution=lognorm,
-                  parameters=[btf.scale_parameter, btf.loc_parameter, btf.scale_parameter])
+                  parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter], median=False)
 
     @staticmethod
     def test_pdfs():
         btf.value_functions(custom_distribution=LogNormalDistribution.scipy_like, scipy_distribution=lognorm,
-                            parameters=[btf.scale_parameter, btf.loc_parameter, btf.scale_parameter], log_check=True)
+                            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter], log_check=True)
 
     @staticmethod
     def test_single_values():
         btf.single_input_n_variables(custom_distribution=LogNormalDistribution.scipy_like,
                                      scipy_distribution=lognorm,
-                                     parameters=[btf.scale_parameter, btf.loc_parameter, btf.scale_parameter],
+                                     parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
                                      log_check=True)
