### 2025-02-07 3622baa
[log-distributions-minor-10]
1. finally correct log_normal in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
