### 2025-02-08 d7680a7
[log-distributions-minor-11]
1. corrected `logNormal_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
