+++ src/tests/test_distributions/test_logNormal_d.py
@@ -1,7 +1,5 @@
 """Created on Dec 14 04:10:02 2024"""
 
-from math import log
-
 import numpy as np
 import pytest
 from scipy.stats import lognorm
@@ -19,7 +17,7 @@ class TestLogNormalDistribution:
     def test_initialization():
         dist = LogNormalDistribution(amplitude=2.0, mu=1.0, std=0.5, normalize=False)
         assert dist.amplitude == 2.0
-        assert dist.mu == log(1)
+        assert dist.mu == np.log(1)
         assert dist.std == 0.5
         assert not dist.norm
 
