### 2025-02-08 00fd864
[log-distributions-minor-12]
1. updated `uniform_d.py` with `log_pdf_` and `log_cdf_` functions
2. updated utilities_d.py`
3. updated `test_scaledInvChiSquare_d.py` and test_uniform_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
