+++ src/pymultifit/distributions/utilities_d.py
@@ -1566,6 +1566,9 @@ def uniform_pdf_(x: np.ndarray,
 
     .. math:: f(x\ |\ a, b) = \dfrac{1}{b-a}
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
@@ -1581,7 +1584,33 @@ def uniform_pdf_(x: np.ndarray,
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
-    return _remove_nans(pdf_)
+    pdf_ = _remove_nans(pdf_)
+
+    return pdf_.item() if scalar_input else pdf_
+
+
+def uniform_log_pdf_(x: fArray,
+                     amplitude: float = 1.0, low: float = 0.0, high: float = 1.0,
+                     normalize: bool = False) -> fArray:
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
+    if x.size == 0:
+        return np.array([])
+
+    high_ = high + low
+    log_pdf_ = np.full(shape=x.shape, fill_value=-np.inf)
+
+    # if high_ == low:
+    #     return log_pdf_
+
+    mask_ = np.logical_and(x >= low, x <= high_)
+    log_pdf_[mask_] = -np.log(high_ - low)
+
+    if not normalize:
+        log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
+
+    return log_pdf_.item() if scalar_input else log_pdf_
 
 
 @doc_inherit(parent=uniform_pdf_, style=doc_style)
@@ -1608,20 +1637,45 @@ def uniform_cdf_(x: np.ndarray,
                         1 &, x > b
                         \end{cases}
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
     high = high + low
+    cdf_ = np.zeros_like(a=x, dtype=float)
 
     if low == high == 0:
         return np.full(shape=x.size, fill_value=np.nan)
 
-    cdf_values = np.zeros_like(a=x, dtype=float)
-    within_bounds = (x >= low) & (x <= high)
-    cdf_values[within_bounds] = (x[within_bounds] - low) / (high - low)
-    cdf_values[x > high] = 1
+    mask_ = (x >= low) & (x <= high)
+    cdf_[mask_] = (x[mask_] - low) / (high - low)
+    cdf_[x > high] = 1
 
-    return cdf_values
+    return cdf_.item() if scalar_input else cdf_
+
+
+def uniform_log_cdf_(x: fArray,
+                     amplitude: float = 1.0, low: float = 0.0, high: float = 1.0,
+                     normalize: bool = False) -> fArray:
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
+    if x.size == 0:
+        return np.array([])
+
+    high = high + low
+    log_cdf_ = np.full(shape=x.shape, fill_value=-np.inf)
+
+    if low == high == 0:
+        return log_cdf_
+
+    mask_ = np.logical_and(x >= low, x <= high)
+    log_cdf_[mask_] = np.log(x[mask_] - low) - np.log(high - low)
+    log_cdf_[x > high] = 0
+
+    return log_cdf_.item() if scalar_input else log_cdf_
 
 
 def scaled_inv_chi_square_pdf_(x, amplitude: float = 1.0, df: float = 1.0, scale: float = 1.0,
[…]
