+++ src/tests/test_distributions/base_test_functions.py
@@ -25,12 +25,16 @@ def edge_cases(distribution, log_check=False):
 
 
 def scaled_distributions(custom_distribution, scipy_distribution, x, parameters, log_check: bool = False,
-                         is_expon=False):
+                         is_expon=False, is_scaled_inv_chi=False):
     dist = custom_distribution(*parameters)
 
     # make an exception for exponential distribution which gets scale = 1/scale
     if is_expon:
         parameters[1] = 1 / parameters[1]
+
+    if is_scaled_inv_chi:
+        parameters[0] = parameters[0] / 2
+        parameters[-1] = parameters[-1] / 2
 
     np.testing.assert_allclose(actual=dist.pdf(x), desired=scipy_distribution(*parameters).pdf(x), rtol=1e-5, atol=1e-8)
     np.testing.assert_allclose(actual=dist.cdf(x), desired=scipy_distribution(*parameters).cdf(x), rtol=1e-5, atol=1e-8)
@@ -41,12 +45,17 @@ def scaled_distributions(custom_distribution, scipy_distribution, x, parameters,
                                    atol=1e-8)
 
 
-def statistics(custom_distribution, scipy_distribution, parameters, mean_variance=False, median=False, is_expon=False):
+def statistics(custom_distribution, scipy_distribution, parameters, mean_variance=False, median=False,
+               is_expon=False, is_scaled_inv_chi=False):
     d_stats = custom_distribution(*parameters).stats()
 
     # make an exception for exponential distribution which gets scale = 1/scale
     if is_expon:
         parameters[1] = 1 / parameters[1]
+
+    if is_scaled_inv_chi:
+        parameters[0] = parameters[0] / 2
+        parameters[-1] = parameters[-1] / 2
 
     if mean_variance:
         scipy_mean, scipy_variance = scipy_distribution.stats(*parameters, moments='mv')
@@ -60,15 +69,18 @@ def statistics(custom_distribution, scipy_distribution, parameters, mean_varianc
         np.testing.assert_allclose(actual=scipy_median, desired=d_stats['median'], rtol=1e-5, atol=1e-8)
 
 
-def stats(custom_distribution, scipy_distribution, parameters, mean_variance=True, median=True, is_expon=False):
+def stats(custom_distribution, scipy_distribution, parameters, mean_variance=True, median=True,
+          is_expon=False, is_scaled_inv_chi=False):
     stack_ = np.column_stack(parameters)
 
     for stack in stack_:
         statistics(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
-                   parameters=stack, mean_variance=mean_variance, median=median, is_expon=is_expon)
+                   parameters=stack, mean_variance=mean_variance, median=median,
+                   is_expon=is_expon, is_scaled_inv_chi=is_scaled_inv_chi)
 
 
-def value_functions(custom_distribution, scipy_distribution, parameters, n_values=50, log_check=False, is_expon=False):
+def value_functions(custom_distribution, scipy_distribution, parameters, n_values=10, log_check=False,
+                    is_expon=False, is_scaled_inv_chi=False):
     x = np.linspace(start=-100, stop=100, num=n_values)
     x = np.concatenate([x, np.array([0, 1])])
     x.sort()
@@ -77,10 +89,11 @@ def value_functions(custom_distribution, scipy_distribution, parameters, n_value
 
     for params in stack_:
         scaled_distributions(custom_distribution=custom_distribution, scipy_distribution=scipy_distribution,
-                             x=x, parameters=params, log_check=log_check, is_expon=is_expon)
+                             x=x, parameters=params, log_check=log_check,
[…]

[…]
