### 2025-02-08 cdfc829
[log-distributions-minor-13]
1. updated `skewNormal_d.py` and `test_skewNormal_d.py`
2. updated `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
