### 2025-02-10 28f6e2b
[log-distributions-minor-14]
1. updated docstring for `utilities_d.py` and `utilities_d.rst`
