+++ src/pymultifit/distributions/utilities_d.py
@@ -27,15 +27,15 @@ from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t, gamma
 from .. import fArray, doc_style
 
 
-def arc_sine_pdf_(x: np.ndarray,
+def arc_sine_pdf_(x: fArray,
                   amplitude: float = 1.0,
-                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.ndarray:
+                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> fArray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
     Parameters
     ----------
-    x : np.ndarray
+    x : fArray
         Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -67,6 +67,9 @@ def arc_sine_pdf_(x: np.ndarray,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
@@ -76,16 +79,18 @@ def arc_sine_pdf_(x: np.ndarray,
 
     pdf_[mask_] = 1 / (np.pi * np.sqrt(y[mask_] * (1 - y[mask_])))
     if not normalize:
-        pdf_[mask_] = _pdf_scaling(pdf_=pdf_[mask_], amplitude=amplitude)
+        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
     pdf_[y == 0] = np.inf
     pdf_[y == 1] = np.inf
 
-    return _remove_nans(pdf_) / scale
+    pdf_ = _remove_nans(pdf_ / scale)
+
+    return pdf_.item() if scalar_input else pdf_
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
-def arc_sine_log_pdf_(x: np.ndarray,
+def arc_sine_log_pdf_(x: fArray,
                       amplitude: float = 1.0,
                       loc: float = 0.0, scale: float = 1.0, normalize: bool = False):
     r"""
@@ -103,6 +108,9 @@ def arc_sine_log_pdf_(x: np.ndarray,
 
     The final logPDF is expressed as :math:`\ell(y) - \ln(\text{scale})`.
     """
+    x = np.asarray(a=x, dtype=float)
+    scalar_input = np.isscalar(x)
+
     if x.size == 0:
         return np.array([])
 
@@ -116,13 +124,15 @@ def arc_sine_log_pdf_(x: np.ndarray,
     log_pdf_[y == 0] = np.inf
     log_pdf_[y == 1] = np.inf
 
-    return log_pdf_ - np.log(scale)
+    log_pdf_ -= np.log(scale)
+
+    return log_pdf_.item() if scalar_input else log_pdf_
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
-def arc_sine_cdf_(x: np.ndarray,
+def arc_sine_cdf_(x: fArray,
                   amplitude: float = 1.0,
-                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False):
+                  loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> fArray:
     r"""
     Compute CDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
@@ -143,6 +153,9 @@ def arc_sine_cdf_(x: np.ndarray,
 
[…]
