### 2025-02-10 d960246
[log-distributions-minor-15]
1. updated docstring for halfNorm pdf, cdf, logpdf, logcdf
