+++ src/pymultifit/distributions/utilities_d.py
@@ -826,7 +826,7 @@ def folded_normal_pdf_(x: fArray,
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
-    pdf_ = _remove_nans(pdf_ / sigma)
+    pdf_ = _remove_nans(pdf_) / sigma
 
     return pdf_.item() if scalar_input else pdf_
 
@@ -1270,7 +1270,7 @@ def half_normal_pdf_(x: fArray,
     Compute PDF for the :class:`~pymultifit.distributions.halfNormal_d.HalfNormalDistribution`.
 
     .. note::
-        The :class:`~pymultifit.distributions.halfNormal_d.HalfNormalDistribution`. is a special case of the
+        The :class:`~pymultifit.distributions.halfNormal_d.HalfNormalDistribution` is a special case of the
         :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution` with :math:`\mu = 0`.
 
     Parameters
@@ -1300,9 +1300,13 @@ def half_normal_pdf_(x: fArray,
     The HalfNormal PDF is defined as:
 
     .. math::
-        f(x\ |\ \sigma) = \sqrt{\dfrac{2}{\pi\sigma^2}}\exp\left[-\dfrac{x^2}{2\sigma^2}\right]
+        f(y\ |\ \sigma) = \sqrt{\dfrac{2}{\pi}}\exp\left(-\dfrac{y^2}{2}\right)
 
-    where :math:`x >= 0`.
+    where :math:`y` is the transformed value of :math:`x`, defined as:
+
+    .. math:: y = \dfrac{x - \text{loc}}{\text{scale}}.
+
+    The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
     x = np.asarray(a=x, dtype=float)
     scalar_input = np.isscalar(x)
@@ -1320,14 +1324,29 @@ def half_normal_pdf_(x: fArray,
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
-    pdf_ = _remove_nans(pdf_ / sigma)
+    pdf_ = _remove_nans(pdf_) / sigma
 
     return pdf_.item() if scalar_input else pdf_
 
-
+@doc_inherit(parent=half_normal_pdf_, style=doc_style)
 def half_normal_log_pdf_(x: fArray,
                          amplitude: float = 1.0, sigma: float = 1.0,
                          loc: float = 0.0, normalize: bool = False) -> fArray:
+    r"""
+    Compute log PDF for the :class:`~pymultifit.distributions.halfNormal_d.HalfNormalDistribution`.
+
+    Notes
+    -----
+    The HalfNormal log PDF is defined as:
+
+    .. math:: \ell(y\ |\ \sigma) = \dfrac{1}{2}\ln\left(\dfrac{2}{\pi}\right) - \dfrac{y^2}{2}
+
+    where :math:`y` is the transformed value of :math:`x`, defined as:
+
+    .. math:: y = \dfrac{x - \text{loc}}{\text{scale}}.
+
+    The final log PDF is expressed as :math:`\ell(y) - \ln\left(\text{scale}\right)`.
+    """
     x = np.asarray(a=x, dtype=float)
     scalar_input = np.isscalar(x)
 
@@ -1365,7 +1384,13 @@ def half_normal_cdf_(x: fArray,
     -----
     The HalfNormal CDF is defined as:
 
-    .. math:: F(x) = \text{erf}\left( \frac{x}{\sqrt{2\sigma^2}}\right)
+    .. math:: F(y) = \text{erf}\left(\frac{y}{\sqrt{2}}\right)
+
[…]
