### 2025-02-10 cd4a458
[log-distributions-major-1]
1. major overhaul of the following functions along with addition of docstrings
    1. `arcsine`
    2. `beta`
    3. `chi2`
    4. `exp`
    5. `foldNorm`
    6. `gammaSR`
    7. `gammaSS`
    6. `gaussian`
