+++ src/pymultifit/distributions/utilities_d.py
@@ -18,11 +18,11 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
            'skew_normal_pdf_', 'skew_normal_cdf_',
            'uniform_pdf_', 'uniform_cdf_', 'uniform_log_pdf_', 'uniform_log_cdf_']
 
-from typing import Union
+from typing import Union, Callable
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t, gammaincc, log_ndtr
+from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr
 
 from .. import fArray, doc_style
 
@@ -53,7 +53,7 @@ def arc_sine_pdf_(x: fArray,
     Returns
     -------
     np.ndarray
-        Array of the same shape as :math:`x`, containing the evaluated PDF values.
+        Array of the same shape as :math:`x`, containing the evaluated values.
 
     Notes
     -----
@@ -67,24 +67,22 @@ def arc_sine_pdf_(x: fArray,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
-    x = np.asarray(a=x, dtype=float)
-    scalar_input = np.isscalar(x)
+    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
 
-    if x.size == 0:
-        return np.array([])
+    if y.size == 0:
+        return y
 
-    y = (x - loc) / scale
-    pdf_ = np.zeros_like(a=y, dtype=float)
     mask_ = np.logical_and(y > 0, y < 1)
 
+    pdf_ = np.zeros_like(a=y, dtype=float)
     pdf_[mask_] = 1 / (np.pi * np.sqrt(y[mask_] * (1 - y[mask_])))
-    if not normalize:
-        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
+    pdf_ = _remove_nans(pdf_) / scale
 
     pdf_[y == 0] = np.inf
     pdf_[y == 1] = np.inf
 
-    pdf_ = _remove_nans(pdf_ / scale)
+    if not normalize:
+        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
     return pdf_.item() if scalar_input else pdf_
 
@@ -108,23 +106,22 @@ def arc_sine_log_pdf_(x: fArray,
 
     The final logPDF is expressed as :math:`\ell(y) - \ln(\text{scale})`.
     """
-    x = np.asarray(a=x, dtype=float)
-    scalar_input = np.isscalar(x)
+    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
 
-    if x.size == 0:
-        return np.array([])
+    if y.size == 0:
+        return y
 
-    y = (x - loc) / scale
-
-    log_pdf_ = np.full_like(a=y, fill_value=-np.inf, dtype=float)
     mask_ = np.logical_and(y > 0, y < 1)
 
+    log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
     log_pdf_[mask_] = -np.log(np.pi) - 0.5 * np.log(y[mask_] - y[mask_]**2)
+    log_pdf_ -= np.log(scale)
 
     log_pdf_[y == 0] = np.inf
     log_pdf_[y == 1] = np.inf
 
-    log_pdf_ -= np.log(scale)
+    if not normalize:
+        log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
 
     return log_pdf_.item() if scalar_input else log_pdf_
 
@@ -153,20 +150,15 @@ def arc_sine_cdf_(x: fArray,
[…]
