+++ docs/source/distributions/utilities_d.rst
@@ -28,12 +28,21 @@ Available for user
 .. autofunction:: exponential_cdf_
 .. autofunction:: exponential_log_cdf_
 .. autofunction:: folded_normal_pdf_
+.. autofunction:: folded_normal_log_pdf_
 .. autofunction:: folded_normal_cdf_
+.. autofunction:: folded_normal_log_cdf_
 .. autofunction:: gamma_sr_pdf_
+.. autofunction:: gamma_sr_log_pdf_
 .. autofunction:: gamma_sr_cdf_
+.. autofunction:: gamma_sr_log_cdf_
 .. autofunction:: gamma_ss_pdf_
+.. autofunction:: gamma_ss_log_pdf_
+.. autofunction:: gamma_ss_cdf_
+.. autofunction:: gamma_ss_log_cdf_
 .. autofunction:: gaussian_pdf_
+.. autofunction:: gaussian_log_pdf_
 .. autofunction:: gaussian_cdf_
+.. autofunction:: gaussian_log_cdf_
 .. autofunction:: half_normal_pdf_
 .. autofunction:: half_normal_log_pdf_
 .. autofunction:: half_normal_cdf_
