### 2025-02-10 bd98521
[log-distributions-minor-16]
1. more cleanup updated in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
