+++ src/pymultifit/distributions/utilities_d.py
@@ -22,7 +22,7 @@ from typing import Union, Callable
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gamma, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr
+from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma
 
 from .. import fArray, doc_style
 
@@ -78,8 +78,9 @@ def arc_sine_pdf_(x: fArray,
     pdf_[mask_] = 1 / (np.pi * np.sqrt(y[mask_] * (1 - y[mask_])))
     pdf_ = _remove_nans(pdf_) / scale
 
-    pdf_[y == 0] = np.inf
-    pdf_[y == 1] = np.inf
+    if not scalar_input:
+        pdf_[y == 0] = np.inf
+        pdf_[y == 1] = np.inf
 
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
@@ -739,16 +740,16 @@ def folded_normal_pdf_(x: fArray,
     where :math:`\phi` is the PDF of :class:`~pymultifit.distributions.gaussian_d.GaussianDistribution`,
     and :math:`y` is the transformed value of :math:`x`, defined as:
 
-    .. math:: y = \dfrac{x - \text{loc}}{\sigma}
+    .. math:: y = \dfrac{x - \text{loc}}{\text{scale}}
 
-    The final PDF is expressed as :math:`f(y)/\sigma`.
+    The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
     x, scalar_input = preprocess_input(x=x)
 
     if x.size == 0:
         return x
 
-    _, pdf_ = _folded(x=x, mean=mean, sigma=sigma, loc=loc, g_func=gaussian_pdf_)
+    _, pdf_ = _folded(x=x, mean=mean, loc=loc, scale=sigma, g_func=gaussian_pdf_)
     pdf_ = _remove_nans(pdf_) / sigma
 
     if not normalize:
@@ -782,7 +783,7 @@ def folded_normal_log_pdf_(x: fArray,
     if x.size == 0:
         return x
 
-    _, pdf_ = _folded(x=x, mean=mean, sigma=sigma, loc=loc, g_func=gaussian_pdf_)
+    _, pdf_ = _folded(x=x, mean=mean, loc=loc, scale=sigma, g_func=gaussian_pdf_)
     log_pdf_ = np.log(pdf_) - np.log(sigma)
 
     if not normalize:
@@ -823,7 +824,7 @@ def folded_normal_cdf_(x: fArray,
     if x.size == 0:
         return x
 
-    mask_, cdf_ = _folded(x=x, mean=mean, sigma=sigma, loc=loc, g_func=gaussian_cdf_)
+    mask_, cdf_ = _folded(x=x, mean=mean, loc=loc, scale=sigma, g_func=gaussian_cdf_)
     cdf_[mask_] -= 1
 
     return cdf_.item() if scalar_input else cdf_
@@ -868,7 +869,7 @@ def folded_normal_log_cdf_(x: fArray,
     return log_cdf_.item() if scalar_input else log_cdf_
 
 
-def _folded(x: fArray, mean: float, sigma: float, loc: float, g_func: Callable):
+def _folded(x: fArray, mean: float, loc: float, scale: float, g_func: Callable):
     r"""
     Precompute the gaussian part of :class:`~pymultifit.distributions.foldedNormal_d.FoldedNormalDistribution`.
 
[…]
