### 2025-02-13 ee66947
[log-distributions-minor-17]
1. updated further docstrings of `pdf_`, `log_pdf_`, `cdf_`, and `log_cdf_` of functions

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
