+++ src/pymultifit/distributions/utilities_d.py
@@ -1,6 +1,6 @@
 """Created on Aug 03 17:13:21 2024"""
 
-__all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
+__all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans', 'preprocess_input',
            'arc_sine_pdf_', 'arc_sine_cdf_', 'arc_sine_log_pdf_', 'arc_sine_log_cdf_',
            'beta_pdf_', 'beta_cdf_', 'beta_log_pdf_', 'beta_log_cdf_',
            'chi_square_pdf_', 'chi_square_cdf_', 'chi_square_log_pdf_', 'chi_square_log_cdf_',
@@ -8,7 +8,7 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans',
            'folded_normal_pdf_', 'folded_normal_cdf_', 'folded_normal_log_pdf_', 'folded_normal_log_cdf_',
            'gamma_sr_pdf_', 'gamma_sr_cdf_', 'gamma_sr_log_pdf_', 'gamma_sr_log_cdf_',
            'gamma_ss_pdf_', 'gamma_ss_log_pdf_', 'gamma_ss_cdf_', 'gamma_ss_log_cdf_',
-           'sym_gen_normal_pdf_', 'sym_gen_normal_cdf_',
+           'sym_gen_normal_pdf_', 'sym_gen_normal_cdf_', 'sym_gen_normal_log_pdf_', 'sym_gen_normal_log_cdf_',
            'gaussian_pdf_', 'gaussian_cdf_', 'gaussian_log_pdf_', 'gaussian_log_cdf_',
            'half_normal_pdf_', 'half_normal_cdf_', 'half_normal_log_pdf_', 'half_normal_log_cdf_',
            'laplace_pdf_', 'laplace_cdf_', 'laplace_log_pdf_', 'laplace_log_cdf_',
@@ -1357,23 +1357,19 @@ def half_normal_pdf_(x: fArray,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
-    x = np.asarray(a=x, dtype=float)
-    scalar_input = np.isscalar(x)
+    y, scalar_input = preprocess_input(x=x, loc=loc, scale=sigma)
 
-    if x.size == 0:
-        return np.array([])
+    if y.size == 0:
+        return y
 
-    y = (x - loc) / sigma
-    pdf_ = np.zeros_like(a=y, dtype=float)
     mask_ = y >= 0
 
-    f1 = np.sqrt(2 / np.pi)
-    pdf_[mask_] = f1 * np.exp(-0.5 * y[mask_]**2)
+    pdf_ = np.zeros_like(a=y, dtype=float)
+    pdf_[mask_] = np.sqrt(2 / np.pi) * np.exp(-0.5 * y[mask_]**2)
+    pdf_ = _remove_nans(pdf_) / sigma
 
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
-
-    pdf_ = _remove_nans(pdf_) / sigma
 
     return pdf_.item() if scalar_input else pdf_
 
@@ -1397,21 +1393,18 @@ def half_normal_log_pdf_(x: fArray,
 
     The final log PDF is expressed as :math:`\ell(y) - \ln\left(\text{scale}\right)`.
     """
-    x = np.asarray(a=x, dtype=float)
-    scalar_input = np.isscalar(x)
+    y, scalar_input = preprocess_input(x=x, loc=loc, scale=sigma)
 
-    if x.size == 0:
-        return np.array([])
+    if y.size == 0:
[…]
