+++ src/tests/test_distributions/test_gamma_sr_d.py
@@ -93,21 +93,22 @@ class TestGammaSSDistribution:
 
     @staticmethod
     def test_edge_cases():
-        btf.edge_cases(distribution=GammaDistributionSR(), log_check=True)
+        btf.edge_cases(distribution=GammaDistributionSS(), log_check=True)
 
     @staticmethod
     def test_stats():
-        btf.stats(custom_distribution=GammaDistributionSR.scipy_like, scipy_distribution=gamma,
+        btf.stats(custom_distribution=GammaDistributionSS.scipy_like, scipy_distribution=gamma,
                   parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter], median=False)
 
     @staticmethod
     def test_pdfs():
-        btf.value_functions(custom_distribution=GammaDistributionSR.scipy_like, scipy_distribution=gamma,
-                            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter], log_check=True)
+        btf.value_functions(custom_distribution=GammaDistributionSS.scipy_like, scipy_distribution=gamma,
+                            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+                            log_check=True)
 
     @staticmethod
     def test_single_values():
-        btf.single_input_n_variables(custom_distribution=GammaDistributionSR.scipy_like,
+        btf.single_input_n_variables(custom_distribution=GammaDistributionSS.scipy_like,
                                      scipy_distribution=gamma,
                                      parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
                                      log_check=True)
