+++ src/pymultifit/distributions/utilities_d.py
@@ -26,6 +26,27 @@ from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, l
 
 from .. import fArray, doc_style
 
+TWO = 2.0
+SQRT_TWO = np.sqrt(TWO)
+LOG_TWO = np.log(TWO)
+LOG_SQRT_TWO = 0.5 * LOG_TWO
+
+PI = np.pi
+SQRT_PI = np.sqrt(PI)
+LOG_PI = np.log(PI)
+LOG_SQRT_PI = 0.5 * LOG_PI
+
+TWO_PI = 2 * PI
+SQRT_TWO_PI = np.sqrt(TWO_PI)
+LOG_TWO_PI = np.log(TWO_PI)
+LOG_SQRT_TWO_PI = 0.5 * LOG_TWO_PI
+
+INV_PI = 1.0 / PI
+TWO_BY_PI = 2.0 * INV_PI
+SQRT_TWO_BY_PI = np.sqrt(TWO_BY_PI)
+LOG_TWO_BY_PI = np.log(TWO_BY_PI)
+LOG_SQRT_TWO_BY_PI = 0.5 * LOG_TWO_BY_PI
+
 
 def arc_sine_pdf_(x: fArray,
                   amplitude: float = 1.0,
@@ -75,7 +96,7 @@ def arc_sine_pdf_(x: fArray,
     mask_ = np.logical_and(y > 0, y < 1)
 
     pdf_ = np.zeros_like(a=y, dtype=float)
-    pdf_[mask_] = 1 / (np.pi * np.sqrt(y[mask_] * (1 - y[mask_])))
+    pdf_[mask_] = 1 / (PI * np.sqrt(y[mask_] * (1 - y[mask_])))
     pdf_ = _remove_nans(pdf_) / scale
 
     if not scalar_input:
@@ -115,7 +136,7 @@ def arc_sine_log_pdf_(x: fArray,
     mask_ = np.logical_and(y > 0, y < 1)
 
     log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
-    log_pdf_[mask_] = -np.log(np.pi) - 0.5 * np.log(y[mask_] - y[mask_]**2)
+    log_pdf_[mask_] = -LOG_PI - 0.5 * np.log(y[mask_] - y[mask_]**2)
     log_pdf_ -= np.log(scale)
 
     log_pdf_[y == 0] = np.inf
@@ -159,7 +180,7 @@ def arc_sine_cdf_(x: fArray,
     mask_ = np.logical_and(y > 0, y < 1)
 
     cdf_ = np.where(y < 1, 0, 1).astype(float)
-    cdf_[mask_] = (2 / np.pi) * np.arcsin(np.sqrt(y[mask_]))
+    cdf_[mask_] = TWO_BY_PI * np.arcsin(np.sqrt(y[mask_]))
 
     return cdf_.item() if scalar_input else cdf_
 
@@ -191,7 +212,7 @@ def arc_sine_log_cdf_(x: fArray,
     mask_ = np.logical_and(y > 0, y < 1)
 
     log_cdf_ = np.where(y < 1, -np.inf, 0)
-    log_cdf_[mask_] = np.log(2 / np.pi) + np.log(np.arcsin(np.sqrt(y[mask_])))
+    log_cdf_[mask_] = LOG_TWO_BY_PI + np.log(np.arcsin(np.sqrt(y[mask_])))
 
     return log_cdf_.item() if scalar_input else log_cdf_
 
@@ -280,9 +301,8 @@ def beta_log_pdf_(x: fArray,
 
     normalization_factor = gammaln(alpha) + gammaln(beta) - gammaln(alpha + beta)
 
[…]
