### 2025-03-07 aae4ed1
[log-corrections-minor-1]
1. corrected laplace_pdf_ and exponential_pdf_ in `utilities_d.py`
2. updated `speed.ipynb`, `summary.ipynb` and `accuracy.ipynb`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
