+++ docs/source/conf.py
@@ -16,7 +16,7 @@ year = datetime.date.today().year
 project = 'pymultifit'
 copyright = f'2024-{year}, Syed Ali Mohsin Bukhari'
 author = 'Syed Ali Mohsin Bukhari'
-release = 'v1.0.4'
+release = 'v1.0.5'
 
 sys.path.insert(0, os.path.abspath('./../../'))
 
@@ -45,12 +45,12 @@ nbsphinx_execute = 'auto'
 source_suffix = {'.rst': 'restructuredtext', '.md': 'restructuredtext'}
 suppress_warnings = ["config.cache"]
 
-# intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
-#                        "numpy": ('https://numpy.org/devdocs', None),
-#                        'scipy': ('https://docs.scipy.org/doc/scipy/', None),
-#                        "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
-#                        "matplotlib": ("https://matplotlib.org/stable/", None),
-#                        "numpydoc": ('https://numpydoc.readthedocs.io/en/latest', None), }
+intersphinx_mapping = {"python": ("https://docs.python.org/3/", None),
+                       "numpy": ('https://numpy.org/devdocs', None),
+                       'scipy': ('https://docs.scipy.org/doc/scipy/', None),
+                       "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
+                       "matplotlib": ("https://matplotlib.org/stable/", None),
+                       "numpydoc": ('https://numpydoc.readthedocs.io/en/latest', None), }
 
 autosummary_generate = True
 autodoc_default_flags = ["members"]
