+++ src/pymultifit/distributions/utilities_d.py
@@ -580,30 +580,19 @@ def exponential_pdf_(x: fArray,
     .. math::
         f(y, \lambda) =
         \begin{cases}
-        \lambda \exp\left[-\lambda y\right] &;& y \geq 0, \\
+        \exp(-y) &;& y \geq 0, \\
         0 &;& y < 0.
         \end{cases}
 
     where, :math:`y` is the transformed value of :math:`x`, defined as:
 
-    .. math:: y = x - \text{loc}
+    .. math:: y = \dfrac{x - \text{loc}}{\theta}
 
-    The final PDF is expressed as :math:`f(y)`.
+    and :math:`\theta = \dfrac{1}{\lambda}`. The final PDF is expressed as :math:`f(y)/\theta`.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc)
-
-    if y.size == 0:
-        return y
-
-    mask_ = y >= 0
-
-    pdf_ = np.zeros(shape=y.shape, dtype=float)
-    pdf_[mask_] = lambda_ * np.exp(-lambda_ * y[mask_])
-
-    if not normalize:
-        pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
-
-    return pdf_.item() if scalar_input else pdf_
+    log_pdf_ = exponential_log_pdf_(x,
+                                    amplitude=amplitude, lambda_=lambda_, loc=loc, normalize=normalize)
+    return np.exp(log_pdf_)
 
 
 @doc_inherit(parent=exponential_pdf_, style=doc_style)
@@ -627,17 +616,18 @@ def exponential_log_pdf_(x: fArray,
     .. math::
         \ell(y, \lambda) =
         \begin{cases}
-        \ln\lambda -\lambda y &;& y \geq 0, \\
-        -inf &;& y < 0.
+        - y &;& y \geq 0, \\
+        -\infty &;& y < 0.
         \end{cases}
 
     where, :math:`y` is the transformed value of :math:`x`, defined as:
 
-    .. math:: y = x - \text{loc}
+    .. math:: y = \dfrac{x - \text{loc}}{\theta}
 
-    The final log PDF is expressed as :math:`\ell(y)`.
+    and :math:`\theta = \dfrac{1}{\lambda}`. The final log PDF is expressed as :math:`\ell(y) - \ln(\theta)`.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc)
+    rate = 1 / lambda_
+    y, scalar_input = preprocess_input(x=x, loc=loc, scale=rate)
 
     if y.size == 0:
         return y
@@ -645,7 +635,7 @@ def exponential_log_pdf_(x: fArray,
     mask_ = y >= 0
 
     log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
-    log_pdf_[mask_] = np.log(lambda_) - (lambda_ * y[mask_])
+    log_pdf_[mask_] = -y[mask_] - np.log(rate)
 
     if not normalize:
         log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
@@ -661,8 +651,8 @@ def exponential_cdf_(x: fArray,
     Compute CDF of :class:`~pymultifit.distributions.exponential_d.ExponentialDistribution`.
 
     .. note::
-        This function uses :func:`~pymultifit.distributions.utilities_d.gamma_sr_cdf_` to calculate the CDF with
-        :math:`\alpha = 1` and :math:`\lambda_\text{gammaSR} = \lambda_\text{expon}`.
+        This function uses :obj:`scipy.special.gammainc` to calculate the CDF with
[…]
