+++ docs/source/index.rst
@@ -54,4 +54,3 @@ With **pyMultiFit**, you can focus more on analyzing your data and less on redef
    tutorials
    api_index
    benchmarks/benchmarks
-   formulae/formulae
