### 2025-03-07 10b784b
[log-corrections-minor-1-1]
1. some updates to tests

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
