### 2025-03-07 89b79aa
[log-corrections-minor-2]
1. updated `mixed_f.py` and `baseFitter.py` for a more robust plotting function
2. added plot_fit in `utilities_f.py`
3. minor updates to `gamma_f.py`
