+++ src/pymultifit/fitters/backend/baseFitter.py
@@ -10,7 +10,7 @@ from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
 from scipy.optimize import Bounds, curve_fit
 
-from ..utilities_f import parameter_logic
+from ..utilities_f import parameter_logic, plot_fit
 from ... import listOfTuplesOrArray, epsilon
 
 
@@ -404,48 +404,10 @@ class BaseFitter:
         else:
             raise ValueError("Either 'mean_values' or 'std_values' must be True.")
 
-    def plot_fit(self, show_individual: bool = False,
-                 x_label: Optional[str] = None, y_label: Optional[str] = None, title: Optional[str] = None,
-                 data_label: Optional[str] = None, axis: Optional[Axes] = None):
-        """
-        Plot the fitted models.
-
-        Parameters
-        ----------
-        show_individual: bool, optional
-            Whether to show individually fitted models or not.
-        x_label: str, optional
-            The label for the x-axis.
-        y_label: str, optional
-            The label for the y-axis.
-        title: str, optional
-            The title for the plot.
-        data_label: str, optional
-            The label for the data.
-        axis: Axes, optional
-            Axes to plot instead of the entire figure. Defaults to None.
-
-        Returns
-        -------
-        plotter
-            The plotter handle for the drawn plot.
-        """
-        if self.params is None:
-            raise RuntimeError("Fit not performed yet. Call fit() first.")
-
-        plotter = plot_xy(x_data=self.x_values, y_data=self.y_values,
-                          data_label=data_label if data_label else 'Data', axis=axis)
-
-        plot_xy(x_data=self.x_values, y_data=self._n_fitter(self.x_values, *self.params),
-                x_label=x_label, y_label=y_label, plot_title=title, data_label='Total Fit',
-                plot_dictionary=LinePlot(color='k'), axis=plotter)
-
-        if show_individual:
-            self._plot_individual_fitter(plotter=plotter)
-
-        plotter.set_xlabel(x_label if x_label else 'X')
-        plotter.set_ylabel(y_label if y_label else 'Y')
-        plotter.set_title(title if title else f'{self.n_fits} {self.__class__.__name__} fit')
-        plt.tight_layout()
-
-        return plotter
+    def plot_fit(self, show_individuals: bool = False,
+                 x_label: Optional[str] = None, y_label: Optional[str] = None, data_label: Union[list[str], str] = None,
+                 title: Optional[str] = None, axis: Optional[Axes] = None):
+        return plot_fit(x_values=self.x_values, y_values=self.y_values, parameters=self.params, n_fits=self.n_fits,
+                        class_name=self.__class__.__name__, _n_fitter=self._n_fitter,
+                        _n_plotter=self._plot_individual_fitter, show_individuals=show_individuals, x_label=x_label,
+                        y_label=y_label, title=title, data_label=data_label, axis=axis)
