### 2025-03-08 #233 @Batalex
Hello @syedalimohsinbukhari, nice to meet you.
I am excited to be part of `pymultifit` review, and I'll get started with my search for reviewers.
