### 2025-03-09 fc603e4
[log-corrections-minor-3]
1. updated the plotting issue in both `baseFitter.py` and `mixed_f.py`
2. minor changes to `conf.py` for version switcher
3. added `versions.json`
4. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
