### 2025-03-09 b74da5c
[log-corrections-minor-3-3]
1. removed `versions.json` will deal with version checker later

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
