+++ docs/source/conf.py
@@ -112,12 +112,6 @@ html_theme_options = {
         "navbar-icon-links"
     ],
     "back_to_top_button": "True",
-    "check_switcher": True,
-    "switcher": {
-        "version_match": "latest",  # This should match the current version
-        # URL to the versions JSON
-        "json_url": "https://github.com/syedalimohsinbukhari/pyMultiFit/blob/log-correction/versions.json",
-    },
 }
 
 html_context = {
