### 2025-03-09 68ed56c
[log-corrections-minor-4]
1. separated `environment.yaml` and `environment[dev].yaml`
2. bumped up pypi version in `version.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
