### 2025-03-09 c139528
[log-corrections-minor-5]
1. updated tutorials

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
