### 2025-03-09 abfd374
[main-minor-14]
1. removed `formulae.rst`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
