### 2025-03-09 #233 @syedalimohsinbukhari
> Hi [@coatless](https://github.com/coatless)
>
> I'm working on some updates and wanted to ask if it'll be okay to push the next version or should I just keep the version as is since I've already mentioned it for submission.

> [@syedalimohsinbukhari](https://github.com/syedalimohsinbukhari) go for it! We're still working on getting editors.

Hi, @coatless, @Batalex

As mentioned previously, I've updated the submission version for `pymultifit` with some streamlining of functionalities without changing much infrastructure. The docs, examples, and tutorails are up to date with this version [v1.0.6](https://github.com/syedalimohsinbukhari/pyMultiFit/releases/tag/v1.0.6).
