### 2025-03-10 6f4ae75
[poly-fitter-minor-1]
1. added `polynomial_f.py`
2. removed `others.py`
3. updated `__init__.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
