### 2025-03-10 #233 @Batalex
Hey, thanks for letting us know. My personal policy when it comes to the reviews I lead as the editor is as follows:
- you can work as you please on the code base, there is no need to freeze anything during the review process. This is important because finding reviewers can take a while, and there is no point in asking you to delay the work multiple times for several months
- we'll ask the reviewers to review the version submitted. Since there is no reviewer at this time, feel free to bump to the latest release.
- however, as soon as reviewers are assigned to the review, the version submitted in the issue header will stay the same. Making sure that reviewers' concern made on an older codebase are properly addressed in the newer one will be your job
