### 2025-03-11 3ec49df
[poly-fitter-minor-2]
1. minor updated for fitter functions
