+++ src/pymultifit/fitters/halfNormal_f.py
@@ -1,6 +1,6 @@
 """Created on Dec 04 23:29:32 2024"""
 
-from math import inf
+import numpy as np
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
@@ -19,8 +19,8 @@ class HalfNormalFitter(BaseFitter):
 
     @staticmethod
     def fit_boundaries():
-        lb = (0, 0, -inf)
-        ub = (inf, inf, inf)
+        lb = (0, 0, -np.inf)
+        ub = (np.inf, np.inf, np.inf)
         return lb, ub
 
     @staticmethod
