+++ src/pymultifit/fitters/foldedNormal_f.py
@@ -1,6 +1,6 @@
 """Created on Dec 04 23:24:55 2024"""
 
-from math import inf
+import numpy as np
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
@@ -17,8 +17,8 @@ class FoldedNormalFitter(BaseFitter):
 
     @staticmethod
     def fit_boundaries():
-        lb = (0, -inf, 0)
-        ub = (inf, inf, inf)
+        lb = (0, -np.inf, 0)
+        ub = (np.inf, np.inf, np.inf)
         return lb, ub
 
     @staticmethod
