+++ src/pymultifit/fitters/logNormal_f.py
@@ -1,6 +1,6 @@
 """Created on Jul 18 19:01:45 2024"""
 
-from math import inf
+import numpy as np
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
@@ -24,8 +24,8 @@ class LogNormalFitter(BaseFitter):
 
     @staticmethod
     def fit_boundaries():
-        lb = (0, -inf, 0, -inf)
-        ub = (inf, inf, inf, inf)
+        lb = (0, -np.inf, 0, -np.inf)
+        ub = (np.inf, np.inf, np.inf, np.inf)
         return lb, ub
 
     @staticmethod
