### 2025-03-13 814b058
[poly-fitter-minor-3]
1. added `line`, `quadratic`, `cube`, `nth_polynomial` functions and classes in `polynomial_f.py` and `utilities_d.py`
2. updated documentation
3. minor changes in distributions
