+++ src/pymultifit/fitters/exponential_f.py
@@ -25,4 +25,4 @@ class ExponentialFitter(BaseFitter):
 
     @staticmethod
     def fitter(x, params):
-        return exponential_pdf_(x, *params, normalize=False)
+        return exponential_pdf_(x, *params)
