+++ src/pymultifit/distributions/backend/__init__.py
@@ -1,4 +1,3 @@
 """Created on Aug 03 22:06:14 2024"""
 
 from .baseDistribution import BaseDistribution
-from .polynomials import *
