+++ src/pymultifit/distributions/generalized/__init__.py
@@ -1,1 +1,4 @@
 """Created on Jan 29 15:42:06 2025"""
+
+from .genNorm_d import SymmetricGeneralizedNormalDistribution
+from .scaledInvChiSquare_d import ScaledInverseChiSquareDistribution
