+++ src/pymultifit/distributions/__init__.py
@@ -9,12 +9,11 @@ from .exponential_d import ExponentialDistribution
 from .foldedNormal_d import FoldedNormalDistribution
 from .gamma_d import GammaDistributionSR, GammaDistributionSS
 from .gaussian_d import GaussianDistribution
-from .generalized.genNorm_d import SymmetricGeneralizedNormalDistribution
+from .generalized import SymmetricGeneralizedNormalDistribution, ScaledInverseChiSquareDistribution
 from .halfNormal_d import HalfNormalDistribution
 from .laplace_d import LaplaceDistribution
 from .logNormal_d import LogNormalDistribution
-from .others import Line
-from .scaledInvChiSquare_d import ScaledInverseChiSquareDistribution
+from .polynomials import Line, Quadratic, Cubic, Polynomial
 from .skewNormal_d import SkewNormalDistribution
 from .uniform_d import UniformDistribution
 
