### 2025-03-14 56aa342
[dev-minor-1]
1. added MPL_COLORS constant to `__init__.py`
2. added data_color and grid to `baseFitter.py` and `mixed_f.py` via `utilities_f.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
