+++ src/pymultifit/fitters/backend/baseFitter.py
@@ -3,15 +3,15 @@
 from itertools import chain
 from typing import Any, Iterable, List, Optional, Tuple, Union
 
-import matplotlib.pyplot as plt
 import numpy as np
+from custom_inherit import doc_inherit
 from matplotlib.axes import Axes
 from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
 from scipy.optimize import Bounds, curve_fit
 
 from ..utilities_f import parameter_logic, _plot_fit
-from ... import listOfTuplesOrArray, epsilon
+from ... import listOfTuplesOrArray, epsilon, MPL_COLORS, doc_style
 
 
 class BaseFitter:
@@ -216,7 +216,7 @@ class BaseFitter:
         """
         x = self.x_values
         params = np.reshape(a=self.params, newshape=(self.n_fits, self.n_par))
-        colors = plt.rcParams['axes.prop_cycle'].by_key()['color'][1:]
+        colors = MPL_COLORS[1:]
         for i, par in enumerate(params):
             color = colors[i % len(colors)]
             plot_xy(x_data=x, y_data=self.fitter(x=x, params=par),
@@ -404,26 +404,13 @@ class BaseFitter:
         else:
             raise ValueError("Either 'mean_values' or 'std_values' must be True.")
 
+    @doc_inherit(parent=_plot_fit, style=doc_style)
     def plot_fit(self, show_individuals: bool = False,
                  x_label: Optional[str] = None, y_label: Optional[str] = None, data_label: Union[list[str], str] = None,
-                 title: Optional[str] = None, axis: Optional[Axes] = None):
+                 title: Optional[str] = None, axis: Optional[Axes] = None,
+                 data_color: str = MPL_COLORS[0], grid: bool = True):
         """
         Plot the fitted models.
-
-        Parameters
-        ----------
-        show_individuals: bool, optional
-            Whether to show individually fitted models or not.
-        x_label: str, optional
-            The label for the x-axis.
-        y_label: str, optional
-            The label for the y-axis.
-        title: str, optional
-            The title for the plot.
-        data_label: str, optional
-            The label for the data.
-        axis: Axes, optional
-            Axes to plot instead of the entire figure. Defaults to None.
 
         Returns
         -------
@@ -433,4 +420,5 @@ class BaseFitter:
         return _plot_fit(x_values=self.x_values, y_values=self.y_values, parameters=self.params, n_fits=self.n_fits,
                          class_name=self.__class__.__name__, _n_fitter=self._n_fitter,
                          _n_plotter=self._plot_individual_fitter, show_individuals=show_individuals, x_label=x_label,
-                         y_label=y_label, title=title, data_label=data_label, axis=axis)
+                         y_label=y_label, title=title, data_label=data_label, axis=axis,
+                         data_color=data_color, grid=grid)
