### 2025-03-18 970076c
[dev-minor-2]
1. updated log_beta in `utilities_d.py`
2. added log_pdf and log_cdf to `skewNormal_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
