+++ src/pymultifit/distributions/utilities_d.py
@@ -15,7 +15,7 @@ __all__ = ['_beta_masking', '_pdf_scaling', '_remove_nans', 'preprocess_input',
            'log_normal_pdf_', 'log_normal_cdf_', 'log_normal_log_pdf_', 'log_normal_log_cdf_',
            'scaled_inv_chi_square_pdf_', 'scaled_inv_chi_square_log_pdf_',
            'scaled_inv_chi_square_cdf_', 'scaled_inv_chi_square_log_cdf_',
-           'skew_normal_pdf_', 'skew_normal_cdf_',
+           'skew_normal_pdf_', 'skew_normal_log_pdf_', 'skew_normal_cdf_', 'skew_normal_log_cdf_',
            'uniform_pdf_', 'uniform_cdf_', 'uniform_log_pdf_', 'uniform_log_cdf_',
            'line', 'linear', 'quadratic', 'cubic', 'nth_polynomial']
 
@@ -23,7 +23,8 @@ from typing import Union, Callable, Optional
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma
+from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma, xlog1py, xlogy
+from scipy.special.cython_special import betaln
 
 from .. import fArray, doc_style
 
@@ -94,11 +95,9 @@ def arc_sine_pdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = np.logical_and(y > 0, y < 1)
-
-    pdf_ = np.zeros_like(a=y, dtype=float)
-    pdf_[mask_] = 1 / (PI * np.sqrt(y[mask_] * (1 - y[mask_])))
-    pdf_ = _remove_nans(pdf_) / scale
+    with np.errstate(divide='ignore', invalid='ignore'):
+        pdf_ = _arcsine(y)
+        pdf_ = _remove_nans(pdf_) / scale
 
     if not scalar_input:
         pdf_[y == 0] = np.inf
@@ -108,6 +107,12 @@ def arc_sine_pdf_(x: fArray,
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
     return pdf_.item() if scalar_input else pdf_
+
+
+def _arcsine(y, un_log=True):
+    pdf = np.where(np.logical_and(y > 0, y < 1),
+                   1 / PI / np.sqrt(y * (1 - y)), 0.0)
+    return pdf if un_log else np.log(pdf)
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
@@ -134,11 +139,8 @@ def arc_sine_log_pdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = np.logical_and(y > 0, y < 1)
-
-    log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
-    log_pdf_[mask_] = -LOG_PI - 0.5 * np.log(y[mask_] - y[mask_]**2)
-    log_pdf_ -= np.log(scale)
+    with np.errstate(divide='ignore', invalid='ignore'):
+        log_pdf_ = _arcsine(y, False) - np.log(scale)
 
     log_pdf_[y == 0] = np.inf
     log_pdf_[y == 1] = np.inf
@@ -266,11 +268,17 @@ def beta_pdf_(x: fArray,
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
[…]
