+++ src/pymultifit/distributions/skewNormal_d.py
@@ -4,7 +4,8 @@ import numpy as np
 
 from .backend import BaseDistribution
 from .backend.errorHandling import NegativeAmplitudeError, NegativeScaleError
-from .utilities_d import skew_normal_cdf_, skew_normal_pdf_
+from .utilities_d import skew_normal_cdf_, skew_normal_pdf_, skew_normal_log_pdf_, skew_normal_log_cdf_
+from .. import fArray
 
 
 class SkewNormalDistribution(BaseDistribution):
@@ -118,10 +119,20 @@ class SkewNormalDistribution(BaseDistribution):
                                 amplitude=self.amplitude, shape=self.shape, loc=self.location, scale=self.scale,
                                 normalize=self.norm)
 
+    def logpdf(self, x: fArray) -> fArray:
+        return skew_normal_log_pdf_(x,
+                                    amplitude=self.amplitude, shape=self.shape, loc=self.location, scale=self.scale,
+                                    normalize=self.norm)
+
     def cdf(self, x):
         return skew_normal_cdf_(x,
                                 amplitude=self.amplitude, shape=self.shape, loc=self.location, scale=self.scale,
                                 normalize=self.norm)
+
+    def logcdf(self, x: fArray) -> fArray:
+        return skew_normal_log_cdf_(x,
+                                    amplitude=self.amplitude, shape=self.shape, loc=self.location, scale=self.scale,
+                                    normalize=self.norm)
 
     def stats(self):
         alpha, omega, epsilon = self.shape, self.scale, self.location
