### 2025-03-18 95396b1
[dev-minor-3]
1. fixed failing tests for arcsine and beta distributions
