### 2025-03-22 cbf5952
[dev-minor-3]
1. added better `frozen` parameter capability in `baseFitter.py`
2. minor changes in `gamma_sr.py` and `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
