### 2025-03-25 555ddc6
[dev-minor-4]
1. `mixed_f.py` is now a subclass for `baseFitter.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
