+++ src/pymultifit/fitters/mixed_f.py
@@ -1,15 +1,13 @@
 """Created on Aug 10 23:08:38 2024"""
 
-import itertools
-from typing import Optional, Tuple, Union, List, Callable
+from itertools import chain
+from typing import Union, List, Callable, Tuple, Iterable
 
 import numpy as np
-from custom_inherit import doc_inherit
-from matplotlib.axes import Axes
 from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
-from scipy.optimize import Bounds, curve_fit
 
+from .backend import BaseFitter
 # importing from files to avoid circular import
 from .chiSquare_f import ChiSquareFitter
 from .exponential_f import ExponentialFitter
@@ -21,9 +19,9 @@ from .laplace_f import LaplaceFitter
 from .logNormal_f import LogNormalFitter
 from .polynomial_f import LineFitter
 from .skewNormal_f import SkewNormalFitter
-from .utilities_f import sanity_check, _plot_fit
-from .. import (epsilon, GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL,
-                GAMMA_SR, GAMMA_SS, NORMAL, HALF_NORMAL, MPL_COLORS, doc_style)
+from .utilities_f import sanity_check
+from .. import (GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL,
+                GAMMA_SR, GAMMA_SS, NORMAL, HALF_NORMAL, epsilon, MPL_COLORS)
 
 # mock initialize the internal classes for auto MixedDataFitter class
 fitter_dict = {CHI_SQUARE: ChiSquareFitter,
@@ -40,7 +38,7 @@ fitter_dict = {CHI_SQUARE: ChiSquareFitter,
                LINE: LineFitter}
 
 
-class MixedDataFitter:
+class MixedDataFitter(BaseFitter):
     r"""
     Class to fit a mixture of different models to data.
 
@@ -53,6 +51,7 @@ class MixedDataFitter:
     def __init__(self, x_values: Union[List, np.ndarray], y_values: Union[List, np.ndarray],
                  model_list: List[str], fitter_dictionary=None, max_iterations: int = 1000):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
+        super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
 
         self.x_values = x_values
         self.y_values = y_values
@@ -61,10 +60,13 @@ class MixedDataFitter:
         self.params = None
         self.covariance = None
 
-        self.fitter_dict = fitter_dictionary or fitter_dict
+        self.p0 = None
 
-        # self._validate_models()
+        self.fitter_dict = fitter_dictionary or fitter_dict
         self.model_function = self._create_model_function()
+
+        # placed in last because the expected parameter count cannot be obtained unless the models are instantiated.
+        self.n_par = self._expected_param_count()
 
     def __repr__(self):
         return (f"{self.__class__.__name__}(x_values={self.x_values}, y_values={self.y_values}, "
@@ -118,30 +120,10 @@ class MixedDataFitter:
 
         return count
 
-    @staticmethod
-    def _format_param(value, t_low=0.001, t_high=10_000) -> str:
-        r"""
-        Formats the parameter value to scientific notation based on its magnitude.
+    def _n_fitter(self, x: np.ndarray, *params: Iterable) -> np.ndarray:
+        return self.model_function(x, *params)
 
-        Parameters
-        ----------
-        value: float
-            The value of the parameter to be formatted.
-        t_low: float, optional
[…]
