### 2025-03-25 f11cb2d
[dev-minor-4-1]
1. fixed the issue with single model parameter not being passed as a tuple
2. added a custom error class in `errorHandling.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
