### 2025-03-26 05a6bad
[dev-minor-5]
1. removed the redundant feature of `scalar_input` to give a single valued output in `utilities_d.py`
