+++ src/pymultifit/distributions/utilities_d.py
@@ -23,8 +23,7 @@ from typing import Union, Callable, Optional
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma, xlog1py, xlogy
-from scipy.special.cython_special import betaln
+from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma
 
 from .. import fArray, doc_style
 
@@ -90,7 +89,7 @@ def arc_sine_pdf_(x: fArray,
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
+    y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
         return y
@@ -99,14 +98,10 @@ def arc_sine_pdf_(x: fArray,
         pdf_ = _arc_sine(y)
         pdf_ = _remove_nans(pdf_) / scale
 
-    if not scalar_input:
-        pdf_[y == 0] = np.inf
-        pdf_[y == 1] = np.inf
-
     if not normalize:
         pdf_ = _pdf_scaling(pdf_=pdf_, amplitude=amplitude)
 
-    return pdf_.item() if scalar_input else pdf_
+    return pdf_
 
 
 def _arc_sine(y: np.ndarray, un_log=True):
@@ -138,7 +133,7 @@ def arc_sine_log_pdf_(x: fArray,
 
     The final logPDF is expressed as :math:`\ell(y) - \ln(\text{scale})`.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
+    y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
         return y
@@ -149,7 +144,7 @@ def arc_sine_log_pdf_(x: fArray,
     if not normalize:
         log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
 
-    return log_pdf_.item() if scalar_input else log_pdf_
+    return log_pdf_
 
 
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
@@ -176,7 +171,7 @@ def arc_sine_cdf_(x: fArray,
 
     .. math:: y = \dfrac{x - \text{loc}}{\text{scale}}.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
+    y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
         return y
@@ -186,7 +181,7 @@ def arc_sine_cdf_(x: fArray,
     cdf_ = np.where(y < 1, 0, 1).astype(float)
     cdf_[mask_] = TWO_BY_PI * np.arcsin(np.sqrt(y[mask_]))
 
-    return cdf_.item() if scalar_input else cdf_
+    return cdf_
 
 
 @doc_inherit(parent=arc_sine_cdf_, style=doc_style)
@@ -208,7 +203,7 @@ def arc_sine_log_cdf_(x: fArray,
 
     The final logCDF is expressed as :math:`\mathcal{L}(y)`.
     """
-    y, scalar_input = preprocess_input(x=x, loc=loc, scale=scale)
+    y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
         return y
@@ -218,7 +213,7 @@ def arc_sine_log_cdf_(x: fArray,
     log_cdf_ = np.where(y < 1, -np.inf, 0)
[…]
