+++ src/pymultifit/distributions/utilities_d.py
@@ -23,7 +23,7 @@ from typing import Union, Callable, Optional
 
 import numpy as np
 from custom_inherit import doc_inherit
-from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma
+from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma, xlogy
 
 from .. import fArray, doc_style
 
@@ -39,14 +39,12 @@ LOG_SQRT_PI = np.log(SQRT_PI)
 
 TWO_PI = 2 * PI
 SQRT_TWO_PI = np.sqrt(TWO_PI)
-LOG_TWO_PI = np.log(TWO_PI)
 LOG_SQRT_TWO_PI = np.log(SQRT_TWO_PI)
 
 INV_PI = 1.0 / PI
 TWO_BY_PI = 2.0 * INV_PI
 SQRT_TWO_BY_PI = np.sqrt(TWO_BY_PI)
 LOG_TWO_BY_PI = np.log(TWO_BY_PI)
-LOG_SQRT_TWO_BY_PI = np.log(SQRT_TWO_BY_PI)
 
 
 def arc_sine_pdf_(x: fArray,
@@ -461,12 +459,10 @@ def chi_square_log_pdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = y > 0
-
-    log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
     df_half = degree_of_freedom / 2
-    log_pdf_[mask_] = (df_half - 1) * np.log(y[mask_]) - (y[mask_] / 2) - (df_half * np.log(2)) - gammaln(df_half)
-    log_pdf_ -= np.log(scale)
+    log_pdf_ = np.where(y >= 0,
+                        xlogy(df_half - 1, y) - (y / 2) - xlogy(df_half, 2) - gammaln(df_half),
+                        -np.inf) - np.log(scale)
 
     if not normalize:
         log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
@@ -505,12 +501,9 @@ def chi_square_cdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = y >= 0
-
-    cdf_ = np.zeros(shape=y.shape, dtype=float)
-    cdf_[mask_] = gammainc(degree_of_freedom / 2, y[mask_] / 2)
-
-    return cdf_
+    return np.where(y >= 0,
+                    gammainc(degree_of_freedom / 2, y / 2),
+                    0.0)
 
 
 @doc_inherit(parent=chi_square_cdf_, style=doc_style)
@@ -662,10 +655,9 @@ def exponential_log_pdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = y >= 0
-
-    log_pdf_ = np.full(shape=y.shape, fill_value=-np.inf)
-    log_pdf_[mask_] = -y[mask_] - np.log(rate)
+    log_pdf_ = np.where(y >= 0,
+                        -y,
+                        -np.inf) - np.log(rate)
 
     if not normalize:
         log_pdf_ = _log_pdf_scaling(log_pdf_=log_pdf_, amplitude=amplitude)
@@ -893,16 +885,13 @@ def folded_normal_log_cdf_(x: fArray,
     if y.size == 0:
         return y
 
-    mask_ = y >= 0
+    q = preprocess_input(y, loc=mean, scale=SQRT_TWO)
+    r = preprocess_input(y, loc=-mean, scale=SQRT_TWO)
 
-    log_cdf_ = np.full(shape=y.shape, fill_value=-np.inf)
-
[…]
