### 2025-04-03 dd13e38
[dev-minor-6]
1. `baseFitter.py`
    1. issue fixed with boundary conditions
    2. added `ci` in `plot_fit` function
    3. added `ci_bounds` function for calculation of CI bounds
2. `mixed_f.py`
    1. added `ci_bounds` - half fixed
3. `utilities_d.py`
    1. fixed laplace functions

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
