+++ src/pymultifit/fitters/backend/baseFitter.py
@@ -1,5 +1,6 @@
 """Created on Jul 18 00:16:01 2024"""
 
+import itertools
 from itertools import chain
 from typing import Any, Iterable, List, Optional, Tuple, Union
 
@@ -266,16 +267,15 @@ class BaseFitter:
         None
             Performs the fitting process with the given constraints.
         """
+        c_name = self.__class__.__name__
         self.p0 = p0
 
         # int -> list converter check
         if isinstance(p0[0], int | float):
             self.p0 = [p0]
 
-        self.n_fits = len(self.p0)
+        self.n_fits = len(self.p0) if c_name != 'MixedDataFitter' else 1
         len_guess = len(list(chain(*self.p0)))
-
-        c_name = self.__class__.__name__
 
         if not c_name == 'MixedDataFitter':
             total_pars = self.n_par * self.n_fits
@@ -287,8 +287,10 @@ class BaseFitter:
         if len(lb) != total_pars and c_name != 'MixedDataFitter':
             self.p0 = self._adjust_parameters(self.p0)
 
-        if ub.shape[0] != len(self.p0[0]) * self.n_fits:
-            raise BoundaryInconsistentWithGuess(f"{ub.shape[0]} != {len(self.p0[0]) * self.n_fits}.")
+        length_ = len(self.p0[0]) if c_name != 'MixedDataFitter' else self._expected_param_count()
+
+        if ub.shape[0] != length_ * self.n_fits:
+            raise BoundaryInconsistentWithGuess(f"{ub.shape[0]} != {length_ * self.n_fits}.")
 
         # remove the Bound parameter for the library to work with previous version of `scipy`
         self.params, self.covariance, *_ = curve_fit(f=self._n_fitter,
@@ -326,7 +328,8 @@ class BaseFitter:
         """
         if self.params is None:
             raise RuntimeError('Fit not performed yet. Call fit() first.')
-        return self._n_fitter(self.x_values, self.params)
+        return self._n_fitter(self.x_values,
+                              *self.params if self.__class__.__name__ == 'MixedDataFitter' else self.params)
 
     def get_model_parameters(self, select: Tuple[int, Any] = None, errors: bool = False):
         r"""
@@ -420,7 +423,7 @@ class BaseFitter:
             raise ValueError("Either 'mean_values' or 'std_values' must be True.")
 
     @doc_inherit(parent=_plot_fit, style=doc_style)
-    def plot_fit(self, show_individuals: bool = False,
+    def plot_fit(self, show_individuals: bool = False, ci: Union[int, list[int]] = None,
                  x_label: Optional[str] = None, y_label: Optional[str] = None, data_label: Union[list[str], str] = None,
                  title: Optional[str] = None, axis: Optional[Axes] = None,
                  data_color: str = MPL_COLORS[0], grid: bool = True):
@@ -432,11 +435,15 @@ class BaseFitter:
         plotter
             The plotter handle for the drawn plot.
         """
-        return _plot_fit(x_values=self.x_values, y_values=self.y_values, parameters=self.params, n_fits=self.n_fits,
+        axes = _plot_fit(x_values=self.x_values, y_values=self.y_values, parameters=self.params, n_fits=self.n_fits,
                          class_name=self.__class__.__name__, _n_fitter=self._n_fitter,
                          _n_plotter=self._plot_individual_fitter, show_individuals=show_individuals, x_label=x_label,
                          y_label=y_label, title=title, data_label=data_label, axis=axis,
                          data_color=data_color, grid=grid)
+        if ci:
+            self.ci_bounds(ci_level=ci, individual_ci=show_individuals, overall_ci=True, plot_it=True, axis=axes)
+
+        return axes
 
[…]
