+++ src/pymultifit/fitters/mixed_f.py
@@ -3,6 +3,7 @@
 from itertools import chain
 from typing import Union, List, Callable, Tuple, Iterable
 
+import matplotlib.pyplot as plt
 import numpy as np
 from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
@@ -20,8 +21,9 @@ from .logNormal_f import LogNormalFitter
 from .polynomial_f import LineFitter
 from .skewNormal_f import SkewNormalFitter
 from .utilities_f import sanity_check
-from .. import (GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE, EXPONENTIAL, FOLDED_NORMAL,
-                GAMMA_SR, GAMMA_SS, NORMAL, HALF_NORMAL, epsilon, MPL_COLORS)
+from .. import (GAUSSIAN, LAPLACE, LINE, LOG_NORMAL, SKEW_NORMAL, CHI_SQUARE,
+                EXPONENTIAL, FOLDED_NORMAL, GAMMA_SR, GAMMA_SS, NORMAL,
+                HALF_NORMAL, epsilon, MPL_COLORS)
 
 # mock initialize the internal classes for auto MixedDataFitter class
 fitter_dict = {CHI_SQUARE: ChiSquareFitter,
@@ -65,7 +67,8 @@ class MixedDataFitter(BaseFitter):
         self.fitter_dict = fitter_dictionary or fitter_dict
         self.model_function = self._create_model_function()
 
-        # placed in last because the expected parameter count cannot be obtained unless the models are instantiated.
+        # placed in last because the expected parameter count cannot be
+        # obtained unless the models are instantiated.
         self.n_par = self._expected_param_count()
 
     def __repr__(self):
@@ -234,3 +237,53 @@ class MixedDataFitter(BaseFitter):
                     ub[sum(par_count[:key - 1]) + (val - 1)] = param_value + epsilon
 
         return lb, ub, list(chain.from_iterable(self.p0))
+
+    def ci_bounds(self, ci_level: Union[int, list[int]] = 3, plot_it: bool = False, overall_ci: bool = False,
+                  individual_ci: bool = False, axis=None):
+        if not overall_ci and not individual_ci:
+            raise ValueError("At least one of `overall_ci` or `individual_ci` must be True.")
+
+        def tuplize_list(lst, sizes):
+            it = iter(lst)
+            return tuple(tuple(next(it) for _ in range(size)) for size in sizes)
+
+        ci_levels = [ci_level] if isinstance(ci_level, int) else ci_level
+        num_samples = 50
+        num_x = len(self.x_values)
+
+        n_par_list = []
+        for model in self.model_list:
+            n_par_list.append(self._instantiate_fitter(model=model, return_values='n_par'))
+
+        mc_iter_values = np.zeros((num_samples, num_x))
+        mc_individual_values = np.zeros((num_samples, self.n_fits, num_x))
+
+        for i in range(num_samples):
+            par_count = 0
+            sampled_params = np.array([np.random.normal(*p)
+                                       for p in self.get_value_error_pair(std_values=True)])
+
+            for model in self.model_list:
+                model_class, n_par = self._instantiate_fitter(model=model, return_values=['class', 'n_par'])
+                individual_fits = model_class.fitter(self.x_values,
+                                                     sampled_params[par_count:par_count + n_par])
+                par_count += n_par
+
+                if overall_ci:
+                    mc_iter_values[i, :] = np.sum(individual_fits, axis=0)
+
+                if individual_ci:
+                    mc_individual_values[i, :, :] = individual_fits
+
+        # results = {}
+        #
+        plt.figure()
+        fitted_curve = self.get_fitted_curve()
[…]
