### 2025-04-04 1edcd62
[dev-minor-7-1]
1. minor changes to `README.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
