### 2025-04-04 9f0134a
[dev-minor-7-2]
1. minor changes to `README.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
