### 2025-04-04 cfc850f
[dev-minor-7-3]
1. minor changes to `README.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
