### 2025-04-07 b515136
[dev-minor-8]
1. minor corrections in `baseFitter.py` for CI
2. moved the ci function from `mixed_f.py` to `baseFitter.py`
