### 2025-04-26 #233 @Batalex
Hi @g4brielvs and @kristinagagalova 👋 Thank you for volunteering to review for pyOpenSci! Please don't hesitate to introduce yourselves.

@syedalimohsinbukhari, I am pleased to report that we have our two reviewers 🎉

- [ ] @g4brielvs survey completed.
- [ ] @kristinagagalova survey completed.

1. Here is the **[reviewers guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html)**. This guide contains all the steps and information needed to complete your review.
2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Please note that we will review the revision `1.06`

Please get in touch with any questions or concerns! Your review is due: May 19th
---

Reviewers: @g4brielvs, @kristinagagalova
Due date: 2025/05/19
