### 2025-05-10 #233 @Batalex
Hey there @syedalimohsinbukhari,
Even as the editor, I like to go through the submissions and give my insights on what could be improved. This is not a formal review and my remarks are more about the style (packaging, code quality) rather than the content (the actual fitters), but I certainly would definitely appreciate it if you implemented some of those suggestions!
I'll probably revisit this if I can find some more time, there is still some content I have not gone through.

# Pyproject.toml

I would advise migrating as much as possible from `setup.py` to `pyproject.toml`.
`pyproject.toml` is now well established in the community, so you can find a lot of documentation for it (shamelessly linking to our own [guide](https://www.pyopensci.org/python-package-guide/tutorials/pyproject-toml.html) on the subject.)

There are several benefits to this approach: [dependencies groups](https://peps.python.org/pep-0735/), having the linter/formatter/test configuration in one single file, using the standard tooling, better readability…
I must admit that I am not super familiar with conda packaging and how we can avoid duplicated information there.

# Automated tooling

I always recommend setting up automation to handle "the boring stuff": code format, linter, type checker, running tests, building docs, etc.

## Formatter

It seems that you already have some kind of IDE built-in formatter for your project (or you spent quite some time yourself aligning those brackets and commas by hand :D).
The problem with this approach is that it is not scalable.
Should an external contributor work on your project, they might not use the same editor.
Therefore, enforcing a consistent format in your codebase will be up to you: that means low added-value commits removing blank spaces or switching single quotes to double ones.

By enforcing a format using a CLI tool such as [`ruff format`](https://docs.astral.sh/ruff/formatter/), you make sure that everyone can apply the same standard code format.
You can add a quality check to your continuous integration on GitHub to avoid adding non-formatted code to your code base.

## Task runner

Building / maintaining a package is hard.
There is only so much we can remember about the various code quality tools we use.
It's even harder when you consider that (1) you write code to solve an issue you have, not just for the sake of it and (2) other contributors might have different experiences when it comes to this stuff.

By using a task runner, you can trivialize a part of this burden by using a third-party tool to delegate the complex commands to run.
The most popular one for Python is [`tox`](https://tox.wiki/en/latest/index.html) but [`nox`](https://nox.thea.codes/en/stable/) has a special place in my heart.

Here is what it can look like. Of course, this specific example is not directly applicable to `pyMultiFit` since it is using `ruff` and `uv`.
`nox` (installed on your system) takes this `noxfile.py` and creates a task for each decorated function. A task is executed in its own virtual environment, so you do not mess with your system or project dependencies at all.

```python
# noxfile.py
import nox

nox.options.default_venv_backend = "uv"
nox.options.reuse_venv = "yes"
nox.options.sessions = ["fmt", "lint"]

@nox.session()
def fmt(session: nox.Session) -> None:
    """Format source code."""
    session.run(
        "uv",
        "sync",
        "--frozen",
        "--only-group",
        "fmt",
        env={"UV_PROJECT_ENVIRONMENT": session.virtualenv.location},
    )
    session.run("ruff", "format", "src", "tests")

@nox.session()
def lint(session: nox.Session) -> None:
    """Lint source code."""
    session.run(
        "uv",
        "sync",
        "--frozen",
        "--group",
        "lint",
        env={"UV_PROJECT_ENVIRONMENT": session.virtualenv.location},
    )
    session.run("ruff", "check", "--fix", "src")
    session.run("ruff", "check", "--fix", "tests")
    session.run("mypy", "src")

@nox.session()
def tests(session: nox.Session) -> None:
    """Run the unit tests."""
    session.run_install(
        "uv",
        "sync",
        "--frozen",
        "--group",
        "unit",
        env={"UV_PROJECT_ENVIRONMENT": session.virtualenv.location},
    )
    session.run("python", "-m", "pytest", *session.posargs)

```

Then, you get this nice user experience out of the box:

```
❯ nox -l
Tasks definition for the nox runner.

Sessions defined in /Users/alex/code/example/noxfile.py:

* fmt -> Format source code.
* lint -> Lint source code.
- tests -> Run the unit tests.

sessions marked with * are selected, sessions marked with - are skipped.
```

Want to format and lint your code? Just run `nox`. Run the unit tests? `nox -s tests`.
You can add as many sessions as you need, like building the docs, fetching an external resource, etc.

Your GitHub workflows are simplified as well:

```yaml
  tests:
    name: Unit test
    runs-on: ubuntu-latest
    steps:
      - name: Checkout
        uses: actions/checkout@v4
      - name: Install nox (and other tools if you need them)
        run: pipx install nox
      - name: Run tests
        run: nox -s tests

```

## Type checker

I commend your efforts on adding some type annotations but while doing so, you must be absolutely certain that they are correct.
My personal take is that it's better to not have any type hint rather than incorrect ones.
For instance, in the gamma fitter example:
```python
import numpy as np
from matplotlib import pyplot as plt

from src.pymultifit import EPSILON
from src.pymultifit.fitters import GammaFitterSR
from src.pymultifit.generators import multi_gamma_sr

params = [(2, 2, 1, 5), (4, 6, 2, 1), (1, 3, 2, 9)]
x = np.linspace(EPSILON, 15, 1000)

noise_level = 0.05
y = multi_gamma_sr(x, params=params, noise_level=noise_level)
```

In the last line, `params=params` is underlined in my editor with the following error, which signals to your user trying out your package that they are doing something wrong:
```
error: Argument of type "list[tuple[int, int, int, int]]" cannot be assigned to parameter "params" of type "listOfTuplesOrArray" in function "multi_gamma_sr"
    [...] (reportArgumentType)
```

The fix is absolutely not trivial. Sure, we could change `listOfTuples` to `List[Tuple[Union[float, int], ...]]`, that way the example provided would still be valid. But what about
```
params = [[2, 2, 1, 5], [4, 6, 2, 1], [1, 3, 2, 9]]
# or
params = np.array([[2.0, 2.0, 1.0, 5.0], [4.0, 6.0, 2.0, 1.0], [1.0, 3.0, 2.0, 9.0]], dtype=np.Float16)
```
?
According to the source code, that should still work, but it will be underlined as well.

Unfortunately, numeric data structures are one of the hardest problems when it comes to type annotations. The main takeaways here are:
- If you want to restrict the type annotations to just a few types, that's fine, users can live with using a list of tuples instead of a 2D np.Float16 array. They just need this restriction explicitly documented and the rest of the code base coherent with this design decision.
- You can be less restrictive with types using something like `params: Sequence[Sequence]` (or even `Any`, because I am not sure whether a 2D numpy array qualifies as a `Sequence[Sequence]`. The typing experience would be looser, but you would defer checking the correctness of the data structure to a runtime check.

To help you with the first point ("the rest of the code base coherent with this design decision"), I encourage you to use a type checker. Yes, another tool to get familiar with :D
The most popular ones for Python are `pyright` and `mypy`.

Be warned that running a type checker for the first time will probably highlight dozens of issues throughout your codebase. It just happens that we make plenty of assumptions while writing code. For instance, in `utilies_f.py`, at the very end `plotter.set_xlabel(x_label if x_label else 'X')` triggers an error for the type checker because `plotter`is either a list or a `matplotlib` Axis. The utility function makes the (probably correct) assumption that we have an axis by this point, and that we can access its method `set_xlabel`, but from a static analysis perspective, we might as well be trying to access this method on a list.

# Misc.

- I encourage you to give the [PEP8 - Style guide](https://peps.python.org/pep-0008/) a read. You sometimes use a format not commonly found in Python codebases (like `listOfTuples` mentioned above. A typical Python codebase would use `ListOfTuples`). Let's be clear, this is not a game changer, it will not make your code run any faster or something like that. However, by enforcing the same stylistic conventions as the ones commonly found in Python projects, you make it easier for people to contribute to your project.
- There are a few instances in the example folder where you pass a `show_individual` argument. The correct one is in plural form.
- I saw a few instances of functions signatures where one of the arguments has a default `None` value. You must be careful with the type annotation, like the `_plot_fit` method with `data_label: Union[list[str], str] = None`. If `data_label` can be either a list of str or a str, then it cannot be `None`, so you should use `Optional` like you did for a few other args.
- There are quite a few variables/modules assigned/imported but never used. It's not a big deal, but it's always nice to cleanup those.
- "Star imports" are usually frown upon in Python (ex: in `distributions/backend/__init__.py` with `from .polynomials import *`). The reason why is that star imports can silently override local definitions, which can lead to hard-to-debug issue. A better approach is to explicitly list what you want to import, or just import the module `import polynomials` to keep the namespace clean.
