### 2025-05-13 258b362
[pyopensci-dev-1]
1. minor corrections in `baseFitter.py`
2. updated `environment.yaml`, `environment[dev].yaml`, `requirements.txt` and `requirements[dev].txt` for `ruff`.
3. created `ruff.toml` (needs working)
4. added ruff to `setup.py`
5. updated version in `version.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
