+++ environment.yaml
@@ -11,3 +11,4 @@ dependencies:
   - pip
   - pip:
       - mpyez
+      - ruff
