### 2025-05-16 beb4dc6
[pyopensci-dev-2]
1. updated `environment.yaml` and `requirements.txt`
2. updated `pyproject.toml`
3. removed `setup.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
