### 2025-05-18 #233 @syedalimohsinbukhari
@Batalex,

Hi, just dropping in here to confirm something.

I've worked ahead of v1.0.6 but couldn't push it before the reviewers were assigned. Those corrections are somewhat significantly different from the version submitted in a few functions.

Keeping in mind that the review is for v1.0.6, I've made a branch that's just a couple of minor corrections ahead of the submitted version, [pyopensci-review](https://github.com/syedalimohsinbukhari/pyMultiFit/tree/pyopensci-review). If it is okay with everyone, I intend to keep the review's correction there before merging into main.

If you have any other suggestions, please let me know.
