### 2025-05-18 #233 @KristinaGagalova
Hi @syedalimohsinbukhari,
Thank you for submitting ```pymultifit``` and working on such a neat project. Please see my review, overall its positive, I only have a few suggestions.

~2
---

The package is thoughtfully designed and well-documented, which reflects best practices in modularity, extensibility, and usability. Please find constructive suggestions aimed at improving the developer and user experience further.

1. Unused normalization parameter in ```exp_pdf```, ```exp_cdf```, and ```Exponential``, but it does not affect the output.
Remove it unless you plan to implement normalization logic.
Example removal:
```
def exp_pdf(x, lambda_):
    return lambda_ * np.exp(-lambda_ * x)
```

2. To prevent overflow in the exponential of the black-body term (especially for large x / kt), in the ```black_body``` function consider capping or regularizing:
```
den_ = np.expm1(np.minimum(x / kt, 700))  # avoid overflow
```

3. As @Batalex suggested previously, it would be good to add a parameter check function for the input and evaluate whether the input is as expected. Its up to the author to decide how strict that should be

Style suggestions:
* It would be suitable to add type hints to functions so that the output is stated clearly
```
def exp_pdf(x: np.ndarray, lambda_: float, normalize: bool = False) -> np.ndarray:
```

Other suggestions:
* It would be valuable to add a simple test with ```pytest``` to ensure that functions, such as ```Exponential().pdf(...)```, returns expected results for known inputs.
* You propose an easy and out-of-the-box support for multifitters, naming```polyfit``` and ```curve_fit```. Do you maybe have any bechnmarking and performance testing on these and how pyMultiFit improves on the mentioned ones?
