### 2025-05-20 3631e6b
[pyopensci-review-1]
1. updated `environment[dev].yaml` and `requirements[dev].txt` to include `ruff`
2. updated `pyproject.toml`
3. removed `setup.py`
4. updated `version.py` to v1.0.7
5. added `ruff.toml`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
