+++ environment[dev].yaml
@@ -30,3 +30,4 @@ dependencies:
   - pip
   - pip:
       - mpyez
+      - ruff
