### 2025-05-25 a672aee
[pyopensci-review-major-1]
1. updated files using ruff formatter manually for now
2. updated ruff.toml

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
