+++ src/pymultifit/distributions/polynomials.py
@@ -1,16 +1,20 @@
 """Created on Aug 10 23:37:54 2024"""
+
 from typing import List
 
 import numpy as np
 
 from .backend import BaseDistribution
 from .utilities_d import line, quadratic, cubic, nth_polynomial
-from .. import fArray
 
 
 class Line(BaseDistribution):
-
-    def __init__(self, slope: float = 1.0, intercept: float = 1.0, normalize: bool = False):
+    def __init__(
+        self,
+        slope: float = 1.0,
+        intercept: float = 1.0,
+        normalize: bool = False,
+    ):
         self.slope = slope
         self.intercept = intercept
 
@@ -29,24 +33,45 @@ class Line(BaseDistribution):
         np.ndarray
             Array of the same shape as :math:`x`, containing the evaluated values.
         """
-        return line(x, slope=self.slope, intercept=self.intercept)
+        return line(
+            x,
+            slope=self.slope,
+            intercept=self.intercept,
+        )
 
 
 class Quadratic(BaseDistribution):
-
-    def __init__(self, a: float = 1.0, b: float = 1.0, c: float = 1.0, normalize: bool = False):
+    def __init__(
+        self,
+        a: float = 1.0,
+        b: float = 1.0,
+        c: float = 1.0,
+        normalize: bool = False,
+    ):
         self.a = a
         self.b = b
         self.c = c
 
         self.norm = normalize
 
-    def pdf(self, x: fArray) -> fArray:
-        return quadratic(x, a=self.a, b=self.b, c=self.c)
+    def pdf(self, x: np.array) -> np.array:
+        return quadratic(
+            x,
+            a=self.a,
+            b=self.b,
+            c=self.c,
+        )
 
 
 class Cubic(BaseDistribution):
-    def __init__(self, a: float = 1.0, b: float = 1.0, c: float = 1.0, d: float = 1.0, normalize: bool = False):
+    def __init__(
+        self,
+        a: float = 1.0,
+        b: float = 1.0,
+        c: float = 1.0,
+        d: float = 1.0,
+        normalize: bool = False,
+    ):
         self.a = a
         self.b = b
         self.c = c
@@ -54,16 +79,24 @@ class Cubic(BaseDistribution):
 
         self.norm = normalize
 
-    def pdf(self, x: fArray) -> fArray:
-        return cubic(x, a=self.a, b=self.b, c=self.c, d=self.d)
+    def pdf(self, x: np.array):
+        return cubic(
+            x,
+            a=self.a,
+            b=self.b,
+            c=self.c,
+            d=self.d,
+        )
 
 
 class Polynomial(BaseDistribution):
-
     def __init__(self, degree: List[int], normalize: bool = False):
         self.degree = degree
 
         self.norm = normalize
 
-    def pdf(self, x: fArray) -> fArray:
-        return nth_polynomial(x, self.degree)
+    def pdf(self, x: np.array):
+        return nth_polynomial(
+            x,
+            coefficients=self.degree,
+        )
