+++ src/pymultifit/fitters/polynomial_f.py
@@ -6,26 +6,38 @@ import numpy as np
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
+from .. import SeqFloat
 from ..distributions.utilities_d import line, nth_polynomial
 
 
 class LineFitter(BaseFitter):
     """A class for fitting a line to the given data."""
 
-    def __init__(self, x_values, y_values, max_iterations: int = 1000):
+    def __init__(
+        self,
+        x_values,
+        y_values,
+        max_iterations: int = 1000,
+    ):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 2
 
     @staticmethod
-    def fitter(x, params: Tuple[float, Any]):
+    def fitter(x: np.ndarray, params: SeqFloat):
         return line(x, *params)
 
 
 class PolynomialFitter(BaseFitter):
     """A class for fitting nth order polynomial to the given data."""
 
-    def __init__(self, x_values, y_values, order, max_iterations=1000):
+    def __init__(
+        self,
+        x_values,
+        y_values,
+        order,
+        max_iterations=1000,
+    ):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = order

[…]
