### 2025-05-25 deea901
[pyopensci-review-minor-2]
1. testing ruff CLI

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
