+++ .github/workflows/pylint.yml
@@ -1,27 +1,0 @@
-name: Pylint
-
-on:
-    push:
-        branches-ignore:
-            - doc
-
-jobs:
-    build:
-        runs-on: ubuntu-latest
-        strategy:
-            matrix:
-                python-version: [ "3.9", "3.10", "3.11" ]
-        steps:
-            -   uses: actions/checkout@v4
-            -   name: Set up Python ${{ matrix.python-version }}
-                uses: actions/setup-python@v4
-                with:
-                    python-version: ${{ matrix.python-version }}
-            -   name: Install dependencies
-                run: |
-                    python -m pip install --upgrade pip
-                    pip install pylint
-                    pip install -r requirements.txt
-            -   name: Analysing the code with pylint
-                run: |
-                    pylint --rcfile=./.pylintrc --fail-under=7 $(git ls-files '*.py')
