+++ .github/workflows/file_linting.yml
@@ -1,0 +1,34 @@
+name: file_linting
+
+on:
+  push:
+    branches-ignore:
+      - doc
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [ "3.9", "3.10", "3.11" ]
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pylint ruff
+          pip install -r requirements.txt
+
+      - name: Analysing the code with Pylint
+        run: |
+          pylint --rcfile=./.pylintrc --fail-under=7 $(git ls-files '*.py')
+
+      - name: Linting the code with Ruff
+        run: |
+          ruff .
