### 2025-05-25 7eebb28
[pyopensci-review-minor-2-1]
1. testing ruff CLI again

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
