+++ ruff.toml
@@ -59,3 +59,5 @@ skip-magic-trailing-comma = false
 
 # Like Black, automatically detect the appropriate line ending.
 line-ending = "auto"
+
+#exclude = ["**"]
