+++ .github/workflows/file_linting.yml
@@ -31,4 +31,4 @@ jobs:
 
       - name: Linting the code with Ruff
         run: |
-          ruff .
+          ruff format ./src/pymultifit/
