### 2025-05-25 d68f5e5
[pyopensci-review-minor-2-2]
1. testing ruff CLI again

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
