### 2025-05-25 6857c45
[pyopensci-review-minor-2-3]
1. testing ruff CLI again

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
