### 2025-05-25 1c30c97
[pyopensci-review-minor-2-4]
1. testing ruff CLI again and pylint

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
