### 2025-05-25 a8e3279
[pyopensci-review-minor-2-5]
1. testing ruff CLI again and pylint

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
