+++ ruff.toml
@@ -60,4 +60,4 @@ skip-magic-trailing-comma = false
 # Like Black, automatically detect the appropriate line ending.
 line-ending = "auto"
 
-#exclude = ["**"]
+exclude = ["**"]
