### 2025-05-25 f14d179
[pyopensci-review-minor-3]
1. automated ruff and pylint with all checks passing

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
