### 2025-05-26 60270e1
[pyopensci-review-minor-4]
1. distributions and generators cleared with pyright and ruff

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
