+++ src/pymultifit/distributions/utilities_d.py
@@ -74,6 +74,7 @@ from typing import Union, Callable
 
 import numpy as np
 from custom_inherit import doc_inherit
+from numpy.typing import NDArray
 from scipy.special import betainc, erf, gammainc, gammaln, owens_t, gammaincc, log_ndtr, ndtr, rgamma, xlogy, gamma
 
 from .. import doc_style
@@ -103,13 +104,13 @@ LOG_TWO_BY_PI = LOG(TWO_BY_PI)
 LOG_SQRT_TWO_BY_PI = 0.5 * LOG_TWO_BY_PI
 
 
-def arc_sine_pdf_(x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> np.array:
+def arc_sine_pdf_(x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> NDArray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
     Parameters
     ----------
-    x : np.array
+    x : NDArray
         Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -162,7 +163,7 @@ def arc_sine_pdf_(x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_log_pdf_(
     x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
-) -> np.array:
+) -> NDArray:
     r"""
     Compute logPDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
@@ -203,7 +204,7 @@ def arc_sine_cdf_(
     loc: float = 0.0,
     scale: float = 1.0,
     normalize: bool = False,
-) -> np.array:
+) -> NDArray:
     y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
@@ -222,7 +223,7 @@ def arc_sine_log_cdf_(
     loc: float = 0.0,
     scale: float = 1.0,
     normalize: bool = False,
-) -> np.array:
+) -> NDArray:
     r"""
     Compute log CDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
 
@@ -257,13 +258,13 @@ def beta_pdf_(
     loc: float = 0.0,
     scale: float = 1.0,
     normalize: bool = False,
-) -> np.array:
+) -> NDArray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Parameters
     ----------
-    x : np.array
+    x : NDArray
         Input array of values where PDF is evaluated.
     amplitude : float, optional
         The amplitude of the PDF. Defaults to 1.0.
@@ -327,7 +328,7 @@ def beta_log_pdf_(
     loc: float = 0.0,
     scale: float = 1.0,
     normalize: bool = False,
-) -> np.array:
+) -> NDArray:
     r"""Compute logPDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Notes
@@ -367,13 +368,13 @@ def beta_cdf_(
     loc: float = 0.0,
     scale: float = 1.0,
     normalize: bool = False,
-) -> np.array:
+) -> NDArray:
     r"""
     Compute CDF for :class:`~pymultifit.distributions.beta_d.BetaDistribution`.
 
     Parameters
     ----------
-    x : np.array
+    x : NDArray
         Input array of values.
     amplitude : float, optional
[…]
