### 2025-05-26 e169a90
[pyopensci-review-minor-5]
1. added `noxfile.py`
2. minor changes to `MANIFEST.in`, `code_coverage.yml` and `publish_pymultifit.yml`
3. fixed LICENSE warning issue with `pyproject.toml`
4. updated `environment[dev].yaml` and `requirements[dev].txt`
5.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
