### 2025-05-26 f33494d
[pyopensci-review-minor-6]
1. minor updates to `utilities_d.py` and fixed gamma functions

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
