+++ src/pymultifit/distributions/utilities_d.py
@@ -70,6 +70,7 @@ __all__ = [
     "nth_polynomial",
 ]
 
+import functools
 from typing import Union, Callable
 
 import numpy as np
@@ -104,6 +105,26 @@ LOG_TWO_BY_PI = LOG(TWO_BY_PI)
 LOG_SQRT_TWO_BY_PI = 0.5 * LOG_TWO_BY_PI
 
 
+def suppress_numpy_warnings(divide="ignore", over="ignore", under="ignore", invalid="ignore"):
+    """
+    A decorator that suppresses NumPy warnings using np.errstate.
+
+    Parameters (all optional):
+        divide, over, under, invalid: Can be 'ignore', 'warn', 'raise', 'call', 'print', or 'log'
+    """
+
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            with np.errstate(divide=divide, over=over, under=under, invalid=invalid):
+                return func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
+
+
+@suppress_numpy_warnings()
 def arc_sine_pdf_(x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False) -> NDArray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
@@ -160,6 +181,7 @@ def arc_sine_pdf_(x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.
     return pdf_
 
 
+@suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_log_pdf_(
     x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
@@ -197,6 +219,7 @@ def arc_sine_log_pdf_(
     return log_pdf_
 
 
+@suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_cdf_(
     x,
@@ -216,6 +239,7 @@ def arc_sine_cdf_(
     return np.select(condlist=[c1, c2], choicelist=[TWO_BY_PI * np.arcsin(np.sqrt(y)), 0.0], default=1.0)
 
 
+@suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_cdf_, style=doc_style)
 def arc_sine_log_cdf_(
     x,
@@ -250,6 +274,7 @@ def arc_sine_log_cdf_(
     return np.select(condlist=[c1, c2], choicelist=[LOG_TWO_BY_PI + LOG(np.arcsin(np.sqrt(y))), 0], default=-INF)
 
 
+@suppress_numpy_warnings()
 def beta_pdf_(
     x,
     amplitude: float = 1.0,
@@ -303,7 +328,6 @@ def beta_pdf_(
 
     The final PDF is expressed as :math:`f(y)/\text{scale}`.
     """
-    # evaluating log_pdf_ for Beta distribution is safer than evaluating direct pdf_ due to less over/under flow issues
     y = preprocess_input(x=x, loc=loc, scale=scale)
 
     if y.size == 0:
@@ -311,7 +335,8 @@ def beta_pdf_(
 
     conditions, main = _beta_expr(y=y, alpha=alpha, beta=beta, un_log=True)
 
[…]
