### 2025-05-30 9987781
[pyopensci-review-minor-7]
1. gamma function updates in all modules

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
