+++ src/pymultifit/generators/__init__.py
@@ -6,8 +6,7 @@ from .generators import (
     multi_base,
     multi_gaussian,
     multi_laplace,
-    multi_gamma_ss,
-    multi_gamma_sr,
+    multi_gamma,
     multi_log_normal,
     multi_skew_normal,
     multi_half_normal,
