+++ src/pymultifit/distributions/arcSine_d.py
@@ -152,7 +152,6 @@ class ArcSineDistribution(BaseDistribution):
         return {
             "mean": mean_,
             "median": median_,
-            "mode": None,
             "variance": variance_,
             "std": np.sqrt(variance_),
         }
