### 2025-06-12 c391263
[pyopensci-review-minor-8]
1. further type setting correction

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
