+++ src/pymultifit/fitters/exponential_f.py
@@ -1,10 +1,11 @@
 """Created on Nov 30 11:30:45 2024"""
 
 import numpy as np
+from numpy.typing import NDArray
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
-from .. import SeqFloat
+from .. import Sequences_, lArray
 from ..distributions.utilities_d import exponential_pdf_
 
 
@@ -13,8 +14,8 @@ class ExponentialFitter(BaseFitter):
 
     def __init__(
         self,
-        x_values,
-        y_values,
+        x_values: lArray,
+        y_values: lArray,
         max_iterations: int = 1000,
     ):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
@@ -30,5 +31,5 @@ class ExponentialFitter(BaseFitter):
         return lb, ub
 
     @staticmethod
-    def fitter(x: np.ndarray, params: SeqFloat):
+    def fitter(x: NDArray, params: Sequences_):
         return exponential_pdf_(x, *params)
