+++ src/pymultifit/fitters/halfNormal_f.py
@@ -1,10 +1,11 @@
 """Created on Dec 04 23:29:32 2024"""
 
 import numpy as np
+from numpy.typing import NDArray
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
-from .. import SeqFloat
+from .. import Sequences_, lArray
 from ..distributions.utilities_d import half_normal_pdf_
 
 
@@ -13,8 +14,8 @@ class HalfNormalFitter(BaseFitter):
 
     def __init__(
         self,
-        x_values,
-        y_values,
+        x_values: lArray,
+        y_values: lArray,
         max_iterations: int = 1000,
     ):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
@@ -30,5 +31,5 @@ class HalfNormalFitter(BaseFitter):
         return lb, ub
 
     @staticmethod
-    def fitter(x: np.ndarray, params: SeqFloat):
+    def fitter(x: NDArray, params: Sequences_):
         return half_normal_pdf_(x, *params)
