### 2025-06-12 bbb2093
[pyopensci-review-minor-9]
1. corrected beta cdf test issues

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
