### 2025-06-20 dae8035
[pyopensci-review-minor-10]
1. removed the dependence of `_remove_nans` function from `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
