### 2025-06-20 09ea2d8
[pyopensci-review-minor-11]
1. major type setting changes
2. added `utilities_d.pyi` for pyright overloading

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
