+++ src/pymultifit/fitters/chiSquare_f.py
@@ -1,11 +1,10 @@
 """Created on Dec 04 03:21:09 2024"""
 
 import numpy as np
-from numpy.typing import NDArray
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
-from .. import Sequences_, lArray
+from .. import lArray, Params_
 from ..distributions.utilities_d import chi_square_pdf_
 
 
@@ -31,5 +30,5 @@ class ChiSquareFitter(BaseFitter):
         return lb, ub
 
     @staticmethod
-    def fitter(x: NDArray, params: Sequences_):
+    def fitter(x: lArray, params: Params_):
         return chi_square_pdf_(x, *params)
