+++ src/pymultifit/fitters/halfNormal_f.py
@@ -1,11 +1,10 @@
 """Created on Dec 04 23:29:32 2024"""
 
 import numpy as np
-from numpy.typing import NDArray
 
 from .backend import BaseFitter
 from .utilities_f import sanity_check
-from .. import Sequences_, lArray
+from .. import lArray, Params_
 from ..distributions.utilities_d import half_normal_pdf_
 
 
@@ -31,5 +30,5 @@ class HalfNormalFitter(BaseFitter):
         return lb, ub
 
     @staticmethod
-    def fitter(x: NDArray, params: Sequences_):
+    def fitter(x: lArray, params: Params_):
         return half_normal_pdf_(x, *params)
