### 2025-06-24 a975ea8
[pyopensci-review-minor-11-1]
1. minor fixes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
