### 2025-06-24 667b801
[pyopensci-review-minor-11-2]
1. updated `utilities_d.pyi` to include missing functions
2. updated `utilities_d.py`
