### 2025-06-24 1a23321
[pyopensci-review-minor-11-3]
1. minor changes to `__init__.py`, `beta_d.py`, `utilities_d.pyi`, `utilities_f.py`
2. type hint settings to `baseFitter.py`
