### 2025-06-24 20ffea4
[pyopensci-review-minor-12]
1. minimized pyright issues throughout the library
