+++ src/pymultifit/fitters/mixed_f.py
@@ -8,6 +8,7 @@ import numpy as np
 from matplotlib.axes import Axes
 from mpyez.backend.uPlotting import LinePlot
 from mpyez.ezPlotting import plot_xy
+from numpy.typing import NDArray
 from scipy.optimize import Bounds, curve_fit
 
 # importing from files to avoid circular import
@@ -34,7 +35,7 @@ from .. import (
     FOLDED_NORMAL,
     GAMMA,
     NORMAL,
-    HALF_NORMAL, Sequences_, lArray,
+    HALF_NORMAL, lArray, Params_,
 )
 
 # mock initialize the internal classes for auto MixedDataFitter class
@@ -73,12 +74,12 @@ class MixedDataFitter:
     ):
         x_values, y_values = sanity_check(x_values=x_values, y_values=y_values)
 
-        self.x_values = x_values
-        self.y_values = y_values
+        self.x_values: NDArray = x_values
+        self.y_values: NDArray = y_values
         self.model_list = model_list
         self.max_iterations = max_iterations
-        self.params = None
-        self.covariance = None
+        self.params: Optional[NDArray] = None
+        self.covariance: Optional[NDArray] = None
 
         self.fitter_dict = fitter_dictionary or fitter_dict
 
@@ -118,7 +119,8 @@ class MixedDataFitter:
             param_index = 0
 
             for model in self.model_list:
-                model_class, n_par = self._instantiate_fitter(model=model, return_values=["class", "n_par"])
+                model_class = self._instantiate_class(model=model)
+                n_par = self._instantiate_n_par(model=model)
                 y += model_class.fitter(x=x, params=params[param_index: param_index + n_par])
                 param_index += n_par
 
@@ -134,8 +136,7 @@ class MixedDataFitter:
         """
         count = 0
         for model in self.model_list:
-            n_par = self._instantiate_fitter(model=model, return_values="n_par")
-            count += n_par
+            count += self._instantiate_n_par(model=model)
 
         return count
 
@@ -162,7 +163,7 @@ class MixedDataFitter:
         """
         return f"{value:.3E}" if t_high < abs(value) or abs(value) < t_low else f"{value:.3f}"
 
-    def _get_bounds(self) -> Tuple[np.ndarray, np.ndarray]:
+    def _get_bounds(self):
         """
         Sets the bounds for each parameter based on the model list.
 
@@ -172,53 +173,27 @@ class MixedDataFitter:
         upper_bounds = []
 
         for model in self.model_list:
-            lb, ub = self._instantiate_fitter(model=model, return_values="bounds")
+            lb, ub = self._instantiate_bounds(model=model)
             lower_bounds.extend(lb)
             upper_bounds.extend(ub)
 
         return np.array(lower_bounds), np.array(upper_bounds)
 
-    def _instantiate_fitter(self, model: str, return_values: Union[str, List[str]] = "class"):
-        """
-        Instantiate the fitter for the specified model and return requested values.
-
-        :param model: The model name as a string.
-        :param return_values: The specific attribute(s) or instance to return.
-                              Options are 'class', 'n_par', and 'bounds'.
-                              Can be a string (for one value) or a list of strings.
-        :return: The requested values as a single value or a tuple.
-        :raises ValueError: If the model is not recognized or return_values are invalid.
-        """
+    def _instantiate_class(self, model: str):
         try:
             fitter_instance = self.fitter_dict[model](x_values=np.array([]), y_values=np.array([]))
         except KeyError:
             raise ValueError(f"Model '{model}' not recognized. Ensure it is defined in the fitter dictionary.")
 
-        valid_options = {"class", "n_par", "bounds"}
-        if isinstance(return_values, str):
-            return_values = [return_values]  # Convert to list for uniform processing
+        return fitter_instance
 
[…]
