### 2025-06-24 5abdee5
[pyopensci-review-minor-12-1]
1. corrections in beta functions
