### 2025-06-24 2b00189
[pyopensci-review-minor-12-2]
1. `__init__.py` fixing
