### 2025-07-01 d0c6479
[pyopensci-review-minor-13]
1. testing negative shape parameter for testing with `test_uniform_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
