### 2025-07-02 d341c42
[pyopensci-review-minor-13-1]
1. reverted the negative scale/shape parameters, it needs more working.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
