+++ src/tests/test_distributions/base_test_functions.py
@@ -5,8 +5,8 @@ import numpy as np
 from ...pymultifit import EPSILON
 
 loc_parameter = np.random.uniform(low=-100, high=100, size=500)
-scale_parameter = np.random.uniform(low=-100, high=100, size=500)
-shape_parameter = np.random.uniform(low=-100, high=100, size=500)
+scale_parameter = np.random.uniform(low=EPSILON, high=100, size=500)
+shape_parameter = np.random.uniform(low=EPSILON, high=100, size=500)
 
 
 def edge_cases(distribution, log_check=False):
