### 2025-07-11 f803cdd
[pyopensci-review-minor-14]
1. opted for `mypy` instaed of `pyright`
2. major changes in docs
3. signature changes in `baseFitter.py`, `utilities_d.py` and child classes
4. added `Deprecated` to primary requirements for deprecation warnings
5. opted for `mypy` instead of `pyright`
6. updated the test suite with the required non-deprecated function
7. updated `noxfile.py` for various amendments and new requirements

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
