+++ src/tests/test_distributions/test_gamma_d.py
@@ -46,18 +46,18 @@ class TestGammaDistribution:
 
     @staticmethod
     def test_stats():
-        btf.stats(custom_distribution=GammaDistribution.scipy_like, scipy_distribution=gamma,
+        btf.stats(custom_distribution=GammaDistribution.from_scipy_params, scipy_distribution=gamma,
                   parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter], median=False)
 
     @staticmethod
     def test_pdfs():
-        btf.value_functions(custom_distribution=GammaDistribution.scipy_like, scipy_distribution=gamma,
+        btf.value_functions(custom_distribution=GammaDistribution.from_scipy_params, scipy_distribution=gamma,
                             parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
                             log_check=True)
 
     @staticmethod
     def test_single_values():
-        btf.single_input_n_variables(custom_distribution=GammaDistribution.scipy_like,
+        btf.single_input_n_variables(custom_distribution=GammaDistribution.from_scipy_params,
                                      scipy_distribution=gamma,
                                      parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
                                      log_check=True)

[…]
