### 2025-07-11 2753da7
[pyopensci-review-minor-14-1]
1. minor pylint optimizaitons

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
