+++ src/pymultifit/distributions/logNormal_d.py
@@ -88,7 +88,7 @@ class LogNormalDistribution(BaseDistribution):
     ):
         if not normalize and amplitude <= 0:
             raise erH.NegativeAmplitudeError()
-        elif std <= 0:
+        if std <= 0:
             raise erH.NegativeStandardDeviationError()
         self.amplitude = 1.0 if normalize else amplitude
         self.mu = np.log(mu)
