+++ src/pymultifit/distributions/gamma_d.py
@@ -96,9 +96,9 @@ class GammaDistribution(BaseDistribution):
     ):
         if not normalize and amplitude <= 0:
             raise erH.NegativeAmplitudeError()
-        elif shape <= 0:
+        if shape <= 0:
             raise erH.NegativeShapeError()
-        elif scale <= 0:
+        if scale <= 0:
             raise erH.NegativeScaleError()
         self.amplitude = 1.0 if normalize else amplitude
         self.shape = shape
