### 2025-07-14 75dfcf7
[pyopensci-review-minor-15]
1. updated benchmarks
2. minor changes in type settings

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
