+++ src/pymultifit/distributions/__init__.py
@@ -3,6 +3,7 @@
 from typing import Optional
 
 from .arcSine_d import ArcSineDistribution
+from .backend import LineFunction, CubicFunction, QuadraticFunction
 from .beta_d import BetaDistribution
 from .chiSquare_d import ChiSquareDistribution
 from .exponential_d import ExponentialDistribution
