### 2025-07-15 c791681
[pyopensci-review-minor-15-1]
1. add a disclaimer for pyopensci-review branch

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
