### 2025-07-15 eb4de48
[pyopensci-review-minor-15-2]
1. added a small patch work to updated requirements using `update_requirements_doc.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
