+++ docs/source/conf.py
@@ -62,7 +62,6 @@ if os.environ.get("READTHEDOCS_VERSION") == "pyopensci-review":
         This is a **review version** of the documentation for PyOpenSci submission.
     """
 
-
 autosummary_generate = True
 autodoc_default_flags = ["members"]
 autodoc_typehints = "signature"
@@ -149,5 +148,11 @@ html_context = {
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
+# def update_installation():
+#     import subprocess
+#     subprocess.run(["python", "update_requirements_doc.py"], check=True)
+
+
 def setup(app):
+    # update_installation()
     app.add_css_file("style.css")
