### 2025-07-15 0f6d9fb
[pyopensci-review-minor-16]
1. updated tutorials
2. updated fitter functions to not give errors on undefined boundaries when directly asked to use instance.fit_boundaries() without defining them.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
