+++ src/pymultifit/fitters/chiSquare_f.py
@@ -23,8 +23,7 @@ class ChiSquareFitter(BaseFitter):
         self.pn_par = 2
         self.sn_par = {"loc": 0.0}
 
-    @staticmethod
-    def fit_boundaries():
+    def fit_boundaries(self):
         lb = (0, 0, -np.inf)
         ub = (np.inf, np.inf, np.inf)
         return lb, ub
