+++ src/pymultifit/fitters/foldedNormal_f.py
@@ -21,10 +21,9 @@ class FoldedNormalFitter(BaseFitter):
         super().__init__(x_values=x_values, y_values=y_values, max_iterations=max_iterations)
         self.n_par = 3
 
-    @staticmethod
-    def fit_boundaries():
-        lb = (0, -np.inf, 0)
-        ub = (np.inf, np.inf, np.inf)
+    def fit_boundaries(self):
+        lb = [0, -np.inf, 0]
+        ub = [np.inf, np.inf, np.inf]
         return lb, ub
 
     @staticmethod
