+++ src/pymultifit/fitters/logNormal_f.py
@@ -28,8 +28,7 @@ class LogNormalFitter(BaseFitter):
         self.pn_par = 3
         self.sn_par = {"loc": 0}
 
-    @staticmethod
-    def fit_boundaries():
+    def fit_boundaries(self):
         lb = (0, -np.inf, 0, -np.inf)
         ub = (np.inf, np.inf, np.inf, np.inf)
         return lb, ub
