### 2025-07-17 #233 @syedalimohsinbukhari
Hi @Batalex,

Just dropping in for updates.

1. I've shifted the project from `setup.py` to `pyproject.toml`.
2. Added and updated `requirements.txt` and `environment.yml` with separate dev dependencies.
3. Incorporated `ruff`, `mypy` and `nox` with `uv` backend.
4. Fixed a ton of type hints and other fixes and further improvements.

Note: The latest docs are currently on `pyopensci-review` branch instead of `latest`

https://pymultifit.readthedocs.io/pyopensci-review/index.html
