### 2025-07-17 #233 @syedalimohsinbukhari
Hi @KristinaGagalova

Starting with your comments,
> Unused normalization parameter in exp_pdf, exp_cdf, and ```Exponential``, but it does not affect the output.
Remove it unless you plan to implement normalization logic.

The normalization logic is baked into the distributions to account for both normalized (area under the curve = 1) and un-normalized distributions. While one could argue that `cdf` doesn’t require normalization, the parameter is retained to maintain consistency across the API for all distribution methods.

> To prevent overflow in the exponential of the black-body term (especially for large x / kt), in the black_body function consider capping or regularizing:

Although this was intended as a user-defined example rather than a core library function, I’ve implemented a decorator to handle overflow/underflow warnings, in line with your suggestion. Internal functions should not give any overflow/underflow warnings.

> As @Batalex suggested previously, it would be good to add a parameter check function for the input and evaluate whether the input is as expected. Its up to the author to decide how strict that should be

Fixed.

> It would be valuable to add a simple test with pytest to ensure that functions, such as Exponential().pdf(...), returns expected results for known inputs.

The test suite is included in src/test/; it is compatible and working with `pytest` and you can also see the report in `codecov` for the `pyopensci-review` branch.

> You propose an easy and out-of-the-box support for multifitters, naming `polyfit` and `curve_fit`. Do you maybe have any bechnmarking and performance testing on these and how pyMultiFit improves on the mentioned ones?

Benchmarking results, including both accuracy and speed tests with a summary, are available at https://pymultifit.readthedocs.io/pyopensci-review/benchmarks/benchmarks.html

**Edit**

The benchmarks are for speed and accuracy against `scipy` functions; as for what makes `pymultifit` an improvement for fitting, it is because of the simpler and modular design that can handle multiple models at the same time. It should be noted that this library uses `curve_fit` in the backend but uses it such that even the models that have different numbers of parameters are handled correctly and without much effort from the user.

The user is facilitated by easy-to-make fitter classes by just defining the function and placing it in the template class. The backend of either `BaseFitter` or `MixedDataFitter` handles the complex tasks depending upon the requested fitter parent class. The facilitations include individual fitting, collective and overall fitting, covariance calculations, and plotting. Plans for confidence interval calculations and more are also there.

The potential of `pymultifit` lies in its ability to internally handle uneven numbers of parameters for different models even when fitting them together. It is also agnostic towards whether the functions are statistical or non-statistical functions, which allows seamless integration of both under a single unified API—supporting fitting and even data generation.

However, `pymultifit` is currently limited to additive model structures, where the total fit is expressed as the sum of individual components. This design decision enables robustness and ease of use, though it may not yet support multiplicative or more complex model compositions.
