### 2025-07-17 d247950
[pyopensci-review-minor-16-2]
1. minor type hint updated

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
