+++ src/pymultifit/distributions/utilities_d.py
@@ -70,7 +70,7 @@ import numpy as np
 import scipy.special as ssp
 from custom_inherit import doc_inherit
 
-from .. import doc_style, LOG, INF
+from .. import doc_style, LOG, INF, ListOrNdArray
 
 TWO = 2.0
 SQRT_TWO = np.sqrt(TWO)
@@ -115,7 +115,7 @@ def suppress_numpy_warnings():
 
 @suppress_numpy_warnings()
 def arc_sine_pdf_(
-    x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
+    x: ListOrNdArray, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
 ) -> np.ndarray:
     r"""
     Compute PDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
@@ -180,7 +180,7 @@ def arc_sine_pdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_log_pdf_(
-    x, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
+    x: ListOrNdArray, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
 ) -> np.ndarray:
     r"""
     Compute logPDF of :class:`~pymultifit.distributions.arcSine_d.ArcSineDistribution`.
@@ -220,7 +220,7 @@ def arc_sine_log_pdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_cdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     loc: float = 0.0,
     scale: float = 1.0,
@@ -243,7 +243,7 @@ def arc_sine_cdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_cdf_, style=doc_style)
 def arc_sine_log_cdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     loc: float = 0.0,
     scale: float = 1.0,
@@ -277,7 +277,7 @@ def arc_sine_log_cdf_(
 
 @suppress_numpy_warnings()
 def beta_pdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     alpha: float = 1.0,
     beta_: float = 1.0,
@@ -349,7 +349,7 @@ def beta_pdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=beta_pdf_, style=doc_style)
 def beta_log_pdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     alpha: float = 1.0,
     beta_: float = 1.0,
@@ -391,7 +391,7 @@ def beta_log_pdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=beta_pdf_, style=doc_style)
 def beta_cdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     alpha: float = 1.0,
     beta_: float = 1.0,
@@ -435,7 +435,7 @@ def beta_cdf_(
 @suppress_numpy_warnings()
 @doc_inherit(parent=beta_cdf_, style=doc_style)
 def beta_log_cdf_(
-    x,
+    x: ListOrNdArray,
     amplitude: float = 1.0,
     alpha: float = 1.0,
     beta_: float = 1.0,
[…]
