### 2025-07-29 #233 @KandiSanjana
Hello @syedalimohsinbukhari, thank you for this submission, this is great! My overall review is good but I see a few things missing in terms of the docs. Here's my review -

---

#### Package Repo
1. I don't see a `CONTRIBUTING.md` in the repo, I'm sure you are going to have people contributing to this, so a good thing to have.
2. Not sure if I missed this, can you add this in too?
`**Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata`

#### README.md
1. Missing badges on the README page although I see a few of these in the docs.
    - Continuous integration and test coverage,
    - A [repostatus.org](https://www.repostatus.org/) badge,
    - Python versions supported
2. Citation infomation in the readme.

#### Paper.md
I don't see a `paper.md` file in the repo containing the following:
- **A short summary** describing the high-level functionality of the software
- **Authors:** A list of authors with their affiliations
- **A statement of need** clearly stating problems the software is designed to solve and its target audience.
- **References:** With DOIs for all those that have one (e.g. papers, datasets, software).
