### 2025-08-13 baf6fc8
[pyopensci-review-minor-16-3]
1. updated `README.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
