### 2025-08-21 4fe78c0
[pyopensci-review-minor-17]
1. added `paper.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
