+++ paper/paper.md
@@ -1,0 +1,244 @@
+---
+title: "pyMultiFit: A Python library for fitting data with multiple models"
+
+tags:
+  - Python
+  - statistics
+  - fitting
+  - Gaussian mixtures
+  - data analysis
+
+author:
+  - name: Syed Ali Mohsin Bukhari
+    orcid: 0009-0002-5888-4163
+    affiliation: ["1", "2"]
+  - name: Iqra Siddique
+    orcid: 0009-0000-2515-6616
+    affiliation: ["3", "4", "5"]
+  - name: Asad Ali
+    orcid: 0000-0002-5164-7942
+    affiliation: ["1", "2"]
+
+affiliation:
+  - name: Department of Applied Mathematics and Statistics, Institute of Space Technology, Islamabad 44000, Pakistan
+    index: 1
+  - name: Space and Astrophysics Research Lab (SARL), National Centre of GIS and Space Applications (NCGSA), Islamabad 44000, Pakistan
+    index: 2
+  - name: Gran Sasso Science Institute (GSSI), Viale Crispi 7, I-67100 L’Aquila, Italy
+    index: 3
+  - name: Dipartimento di Fisica, Università di Trento, Via Sommarive 14 I-38123 Trento, Italy
+    index: 4
+  - name: INFN Laboratori Nazionali del Gran Sasso, Via Acitelli 22, I-67100 Assergi, L’Aquila, Italy
+    index: 5
+
+date: 21 August 2025
+
+bibliography: paper.bib
+---
+
+
+# Summary
+
+`pyMultiFit` is an open-source Python library that streamlines multi-model curve fitting for scientific applications.
+It provides a uniform API for both statistical and non-statistical models, with built-in support for common
+distributions such as Gaussian, Laplace, SkewedNormal, LogNormal, and Exponential.
+The library emphasizes extensibility through a lightweight `BaseFitter` interface, enabling users to define custom
+models with minimal boilerplate.
+In addition to fitting, `pyMultiFit` includes utilities for synthetic data generation and integrated statistical tools for
+constructing synthetic datasets.
+
+[//]: # (While most built-in fitters target homogeneous data, `pyMultiFit` also includes a `MixedDataFitter` that enables the simultaneous fitting of heterogeneous models &#40;e.g., Gaussian + SkewNormal + Linear&#41;, making it especially useful for applications in signal analysis, spectral modeling, and other multi-component workflows.)
+
+Together, these features make `pyMultiFit` especially well-suited for researchers working with **signals**, **spectra**,
+or **experimental datasets** where multiple overlapping components are common.
+
+# Statement of Need
+
+Data fitting is a cornerstone of experimental and analytical workflows across the sciences.
+Yet, widely used scientific libraries such as `numpy`[@harris2020array] and `scipy` [@virtanen2020scipy] provide only
+limited functionality for fitting **multiple models or model mixtures** concurrently.
+Researchers working with multi-component fits—common in domains such as signal analysis, spectroscopy, and high-energy
+physics—are often required to implement complex and repetitive boilerplate code to manage these workflows.
+
+Several modern libraries have been developed to extend beyond these limitations.
+For example, `zfit` [@zfit2019] leverages `tensorflow` [@tensorflow2016] to support high-performance statistical
+fitting, `pyAutoFit` [@pyautofit2021] facilitates hierarchical model construction, and specialized
+`Mixture-Models` [@MM2024] packages provide probabilistic mixture fitting.
+While powerful, these tools are either tightly bound to heavy backends, optimized for niche applications, or difficult
+to adapt for general multimodel curve-fitting tasks.
+
[…]
