### 2025-08-21 f3d3c52
[pyopensci-review-minor-17-1]
1. renamed `distribution_test.py` to `distribution_.py` to avoid issue with codecov

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
