### 2025-08-21 #233 @syedalimohsinbukhari
hi @KandiSanjana

> **Package Repo**
> 1. I don't see a CONTRIBUTING.md in the repo, I'm sure you are going to have people contributing to this, so a good thing to have.

Done.

> 2. Not sure if I missed this, can you add this in too?
> `**Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata`

This information is already available in `pyproject.toml`

> **README.md**
> 1. Missing badges on the README page although I see a few of these in the docs.
a. Continuous integration and test coverage,
b. A [repostatus.org](https://www.repostatus.org/) badge,
c. Python versions supported
> 2. Citation infomation in the readme.

Added CI/CV + repostatus + python version badges. But wouldn't citation be after the paper.md is accepted?

> Paper.md
> I don't see a paper.md file in the repo containing the following:
> - **A short summary** describing the high-level functionality of the software
> - **Authors:** A list of authors with their affiliations
> - **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> - **References:** With DOIs for all those that have one (e.g. papers, datasets, software).

Added
