### 2025-08-21 a325c7c
[pyopensci-review-minor-17-2]
1. added `CONTRIBUTING.md`
2. updated `README.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
