### 2025-10-05 #233 @Batalex
---

Then, it's my great pleasure to announce that...

🎉  `pymultifit` has been approved by pyOpenSci! Thank you @syedalimohsinbukhari for submitting `pymultifit` and many thanks to @KandiSanjana and @KristinaGagalova for reviewing this package! 😸
It took a while, but we got there!

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
