### 2025-10-05 e6a0ad6
[pyopensci-review-major-2]
1. finalized `pyproject.toml` for publishing `v1.0.7` - reviews completed.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
