### 2025-10-05 dda0bab
[main-minor-15]
1. updated `pyproject.toml`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
