+++ pyproject.toml
@@ -14,16 +14,16 @@ dynamic = ["version"]
 dependencies = ["numpy<2.1.0", "matplotlib", "scipy", "custom-inherit", "mpyez"]
 keywords = ["fitting", "multi-modal", "gaussian", "lognormal", "skew", "curve fitting", "scipy"]
 classifiers = [
-    "Classifier: Development Status :: 5 - Production/Stable",
-    "Classifier: Intended Audience :: Science/Research",
-    "Classifier: License :: OSI Approved :: MIT License",
-    "Classifier: Operating System :: OS Independent",
-    "Classifier: Programming Language :: Python :: 3.9",
-    "Classifier: Programming Language :: Python :: 3.10",
-    "Classifier: Programming Language :: Python :: 3.11",
-    "Classifier: Topic :: Scientific/Engineering :: Astronomy",
-    "Classifier: Topic :: Scientific/Engineering :: Mathematics",
-    "Classifier: Topic :: Scientific/Engineering :: Physics"
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Scientific/Engineering :: Astronomy",
+    "Topic :: Scientific/Engineering :: Mathematics",
+    "Topic :: Scientific/Engineering :: Physics"
 ]
 
 [tool.setuptools.dynamic]
