### 2025-10-05 d34a6d8
[main-minor-16]
1. updated `README.md` with badges

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
