### 2025-10-05 #233 @syedalimohsinbukhari
## Author Wrap Up Tasks

* [x] Add the badge for pyOpenSci peer-review to the README.md of . The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

Hi @Batalex, done with the wrap up tasks. I'm curious about how to proceed towards JOSS since I've already made a paper.md for it as well.
