### 2025-10-16 27cba9f
[main-minor-16-1]
1. testing `draft-pdf.yml` for JOSS

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
