+++ .github/workflows/draft-pdf.yml
@@ -1,0 +1,28 @@
+name: Draft PDF
+on:
+  push:
+    branches:
+      - main
+    paths:
+      - inst/**
+      - .github/workflows/draft-pdf.yml
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Paper Draft
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@v1
+        with:
+          journal: joss
+          paper-path: inst/paper.md
+
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          path: inst/paper.pdf
