### 2025-10-16 a58f645
[main-minor-16-2]
1. minor changes to `draft-pdf.yml`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
