+++ .github/workflows/draft-pdf.yml
@@ -16,7 +16,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Build draft PDF
-        uses: openjournals/openjournals-draft-action@v1
+        uses: openjournals/openjournals-draft-action@main
         with:
           journal: joss
           paper-path: inst/paper.md
