### 2025-10-16 4314e3c
[main-minor-16-3]
1. minor changes to `draft-pdf.yml`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
