### 2025-10-16 269c3db
[main-minor-17]
1. changed `paper.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
