+++ inst/paper.md
@@ -1,24 +1,22 @@
 ---
-title: "pyMultiFit: A Python library for fitting data with multiple models"
-
+title: 'pyMultiFit: A Python library for fitting data with multiple models'
 tags:
   - Python
   - statistics
   - fitting
   - Gaussian mixtures
   - data analysis
-
-author:
+authors:
   - name: Syed Ali Mohsin Bukhari
     orcid: 0009-0002-5888-4163
-    affiliation: ["1", "2"]
+    corresponding: true
+    affiliation: "1, 2"
   - name: Iqra Siddique
     orcid: 0009-0000-2515-6616
-    affiliation: ["3", "4", "5"]
+    affiliation: "3, 4, 5"
   - name: Asad Ali
     orcid: 0000-0002-5164-7942
-    affiliation: ["1", "2"]
-
+    affiliation: "1, 2"
 affiliation:
   - name: Department of Applied Mathematics and Statistics, Institute of Space Technology, Islamabad 44000, Pakistan
     index: 1
@@ -30,12 +28,9 @@ affiliation:
     index: 4
   - name: INFN Laboratori Nazionali del Gran Sasso, Via Acitelli 22, I-67100 Assergi, L’Aquila, Italy
     index: 5
-
 date: 21 August 2025
-
 bibliography: paper.bib
 ---
-
 
 # Summary
 
@@ -55,7 +50,7 @@ or **experimental datasets** where multiple overlapping components are common.
 # Statement of Need
 
 Data fitting is a cornerstone of experimental and analytical workflows across the sciences.
-Yet, widely used scientific libraries such as `numpy`[@harris2020array] and `scipy` [@virtanen2020scipy] provide only
+Yet, widely used scientific libraries such as `numpy` [@harris2020array] and `scipy` [@virtanen2020scipy] provide only
 limited functionality for fitting **multiple models or model mixtures** concurrently.
 Researchers working with multi-component fits—common in domains such as signal analysis, spectroscopy, and high-energy
 physics—are often required to implement complex and repetitive boilerplate code to manage these workflows.
@@ -125,27 +120,19 @@ custom_with_scipy = GaussianDistribution.from_scipy_params(loc=-1,
 f, ax = plt.subplots(1, 2, figsize=(15, 5))
 ax[0].plot(x, custom_distribution, 
            label='custom distribution\nw/o\nscipy parametrization')
-ax[0].plot(x, scipy_distribution, 
-           label='scipy distribution')
-ax[0].set_xlabel('x')
-ax[0].set_ylabel('pdf')
-ax[0].set_title('Gaussian Distribution')
-ax[0].grid(True, alpha=0.5, ls='--')
-ax[0].legend(loc='best')
+ax[0].plot(x, scipy_distribution, label='scipy distribution')
 
 ax[1].plot(x, custom_with_scipy, 
            label='custom distribution\nw\nscipy parametrization')
-ax[1].plot(x, scipy_distribution, 
-           label='scipy distribution')
-ax[1].set_xlabel('x')
-ax[1].set_ylabel('pdf')
-ax[1].set_title('Gaussian Distribution')
-ax[1].grid(True, alpha=0.5, ls='--')
-ax[1].legend(loc='best')
+ax[1].plot(x, scipy_distribution, label='scipy distribution')
+
+[i.set_xlabel('x') for i in ax]
+[i.set_ylabel('pdf') for i in ax]
+[i.set_title('Gaussian Distribution') for i in ax]
+[i.grid(True, alpha=0.5, ls='--') for i in ax]
+[i.legend(loc='best') for i in ax]
 
 plt.tight_layout()
[…]
