### 2025-10-16 74160c4
[main-minor-17-1]
1. changed `paper.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
