### 2025-10-19 dde8377
[main-minor-17-2]
1. updated affiliations in `paper.md`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
