### 2025-10-23 f618978
Update figures and section headings in paper.md
