### 2025-10-24 69bd33c
Remove figure captions from markdown images
