### 2025-10-24 c3d9b71
[main-major-23]
1. JOSS acceptance

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
