+++ CITATION.cff
@@ -1,0 +1,44 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Bukhari
+  given-names: Syed Ali Mohsin
+  orcid: "https://orcid.org/0009-0002-5888-4163"
+- family-names: Siddique
+  given-names: Iqra
+  orcid: "https://orcid.org/0009-0000-2515-6616"
+- family-names: Ali
+  given-names: Asad
+  orcid: "https://orcid.org/0000-0002-5164-7942"
+contact:
+- family-names: Bukhari
+  given-names: Syed Ali Mohsin
+  orcid: "https://orcid.org/0009-0002-5888-4163"
+doi: 10.5281/zenodo.17273507
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Bukhari
+    given-names: Syed Ali Mohsin
+    orcid: "https://orcid.org/0009-0002-5888-4163"
+  - family-names: Siddique
+    given-names: Iqra
+    orcid: "https://orcid.org/0009-0000-2515-6616"
+  - family-names: Ali
+    given-names: Asad
+    orcid: "https://orcid.org/0000-0002-5164-7942"
+  date-published: 2025-10-24
+  doi: 10.21105/joss.09295
+  issn: 2475-9066
+  issue: 114
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9295
+  title: "pyMultiFit: A Python library for fitting data with multiple
+    models"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09295"
+  volume: 10
+title: "pyMultiFit: A Python library for fitting data with multiple
+  models"
