### 2025-10-24 80bdd14
[v108-minor-1]
1. replaced ruff with black and formatted the files
2. updated `pyproject.toml`, `environment[dev].yaml`, and `requirements[dev].txt`
3. updated `functions.py` in benchmarks
4. updated to v1.0.8 ->

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
