+++ src/tests/test_distributions/test_gen_norm.py
@@ -58,32 +58,40 @@ class TestSymNormalDistribution:
             d_stats = _distribution.stats()
 
             # Scipy calculations
-            scipy_mean, scipy_variance = gennorm.stats(beta=shape, loc=loc, scale=scale, moments='mv')
+            scipy_mean, scipy_variance = gennorm.stats(beta=shape, loc=loc, scale=scale, moments="mv")
             scipy_median = gennorm.median(beta=shape, loc=loc, scale=scale)
             scipy_stddev = np.sqrt(scipy_variance)
 
             # Assertions for mean and variance
-            np.testing.assert_allclose(actual=scipy_mean, desired=d_stats['mean'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_variance, desired=d_stats['variance'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_median, desired=d_stats['median'], rtol=1e-5, atol=1e-8)
-            np.testing.assert_allclose(actual=scipy_stddev, desired=d_stats['std'], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_mean, desired=d_stats["mean"], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_variance, desired=d_stats["variance"], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_median, desired=d_stats["median"], rtol=1e-5, atol=1e-8)
+            np.testing.assert_allclose(actual=scipy_stddev, desired=d_stats["std"], rtol=1e-5, atol=1e-8)
 
     @staticmethod
     def test_pdf_cdf():
-        def _cdf_pdf_custom(x_, dist_, what='cdf'):
-            return dist_.cdf(x_) if what == 'cdf' else dist_.pdf(x_)
+        def _cdf_pdf_custom(x_, dist_, what="cdf"):
+            return dist_.cdf(x_) if what == "cdf" else dist_.pdf(x_)
 
-        def _cdf_pdf_scipy(x_, shape, loc, scale, what='cdf'):
-            return [gennorm.cdf(x_, beta=shape, loc=loc, scale=scale) if what == 'cdf' else
-                    gennorm.pdf(x_, beta=shape, loc=loc, scale=scale)][0]
+        def _cdf_pdf_scipy(x_, shape, loc, scale, what="cdf"):
+            return [
+                (
+                    gennorm.cdf(x_, beta=shape, loc=loc, scale=scale)
+                    if what == "cdf"
+                    else gennorm.pdf(x_, beta=shape, loc=loc, scale=scale)
+                )
+            ][0]
 
-        for i in ['cdf', 'pdf']:
+        for i in ["cdf", "pdf"]:
             for _ in range(500):  # Run 50 random tests
                 shape_ = np.random.uniform(low=EPSILON, high=100)
                 loc_ = np.random.uniform(low=-100, high=100)
                 scale_ = np.random.uniform(low=EPSILON, high=100)
                 x = np.linspace(start=loc_ - 10, stop=loc_ + 10, num=50)
-                distribution = SymmetricGeneralizedNormalDistribution.from_scipy_params(beta=shape_, loc=loc_, scale=scale_)
+                distribution = SymmetricGeneralizedNormalDistribution.from_scipy_params(
+                    beta=shape_, loc=loc_, scale=scale_
+                )
                 expected = _cdf_pdf_scipy(x_=x, shape=shape_, loc=loc_, scale=scale_, what=i)
-                np.testing.assert_allclose(actual=_cdf_pdf_custom(x_=x, dist_=distribution, what=i),
[…]
