+++ src/tests/test_distributions/test_skewNormal_d.py
@@ -42,18 +42,25 @@ class TestSkewNormalDistribution:
 
     @staticmethod
     def test_stats():
-        btf.stats(custom_distribution=SkewNormalDistribution.from_scipy_params, scipy_distribution=skewnorm,
-                  parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
-                  median=False)
+        btf.stats(
+            custom_distribution=SkewNormalDistribution.from_scipy_params,
+            scipy_distribution=skewnorm,
+            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            median=False,
+        )
 
     @staticmethod
     def test_pdfs():
-        btf.value_functions(custom_distribution=SkewNormalDistribution.from_scipy_params,
-                            scipy_distribution=skewnorm,
-                            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter])
+        btf.value_functions(
+            custom_distribution=SkewNormalDistribution.from_scipy_params,
+            scipy_distribution=skewnorm,
+            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+        )
 
     @staticmethod
     def test_single_values():
-        btf.single_input_n_variables(custom_distribution=SkewNormalDistribution.from_scipy_params,
-                                     scipy_distribution=skewnorm,
-                                     parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter])
+        btf.single_input_n_variables(
+            custom_distribution=SkewNormalDistribution.from_scipy_params,
+            scipy_distribution=skewnorm,
+            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+        )
