### 2025-10-25 2b7b44d
[v110-minor-1-2]
1. updated `file_linting.yml` with black installation

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
