+++ .github/workflows/file_linting.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pylint ruff
+          pip install pylint black
           pip install -r requirements.txt
 
       - name: Analysing the code with Pylint
