### 2025-10-26 5fd99b5
[v110-minor-2]
1. updated benchmark and examples
2. updated `environment[dev].yaml` for more conda-centric setup

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
