+++ benchmarks/functions.py
@@ -31,7 +31,7 @@ def compare_accuracy(x, custom_dist, scipy_dist):
     pdf_scipy = scipy_dist.pdf(x)
     cdf_scipy = scipy_dist.cdf(x)
 
-    with np.errstate(invalid='ignore'):
+    with np.errstate(invalid="ignore"):
         pdf_abs_diff = np.nan_to_num(np.abs(pdf_custom - pdf_scipy), False, 0) + EPSILON
         cdf_abs_diff = np.nan_to_num(np.abs(cdf_custom - cdf_scipy), False, 0) + EPSILON
 
@@ -39,16 +39,16 @@ def compare_accuracy(x, custom_dist, scipy_dist):
     logcdf_custom = custom_dist.logcdf(x)
     logpdf_scipy = scipy_dist.logpdf(x)
     logcdf_scipy = scipy_dist.logcdf(x)
-    with np.errstate(invalid='ignore'):
-        logpdf_abs_diff = np.nan_to_num(np.abs(logpdf_custom - logpdf_scipy),
-                                        False, 0) + EPSILON
-        logcdf_abs_diff = np.nan_to_num(np.abs(logcdf_custom - logcdf_scipy),
-                                        False, 0) + EPSILON
+    with np.errstate(invalid="ignore"):
+        logpdf_abs_diff = np.nan_to_num(np.abs(logpdf_custom - logpdf_scipy), False, 0) + EPSILON
+        logcdf_abs_diff = np.nan_to_num(np.abs(logcdf_custom - logcdf_scipy), False, 0) + EPSILON
 
-    return {"pdf_abs_diff": pdf_abs_diff,
-            "log_pdf_abs_diff": logpdf_abs_diff,
-            "cdf_abs_diff": cdf_abs_diff,
-            "log_cdf_abs_diff": logcdf_abs_diff}
+    return {
+        "pdf_abs_diff": pdf_abs_diff,
+        "log_pdf_abs_diff": logpdf_abs_diff,
+        "cdf_abs_diff": cdf_abs_diff,
+        "log_cdf_abs_diff": logcdf_abs_diff,
+    }
 
 
 # Plotting Function
@@ -56,7 +56,7 @@ def plot_accuracy(x, results, title_suffix):
     plt.figure(figsize=(12, 8))
 
     plt.subplot(2, 2, 1)
-    plt.plot(x, results["pdf_abs_diff"], label="PDF Absolute Diff", marker='.')
+    plt.plot(x, results["pdf_abs_diff"], label="PDF Absolute Diff", marker=".")
     plt.xscale("log")
     plt.yscale("log")
     plt.xlabel("x")
@@ -68,7 +68,7 @@ def plot_accuracy(x, results, title_suffix):
 
     # if results.get('log_pdf_abs_diff') is not None:
     plt.subplot(2, 2, 2)
-    plt.plot(x, results["log_pdf_abs_diff"], label="log PDF Absolute Diff", marker='.')
+    plt.plot(x, results["log_pdf_abs_diff"], label="log PDF Absolute Diff", marker=".")
     plt.xscale("log")
     plt.yscale("log")
     plt.xlabel("x")
@@ -79,7 +79,7 @@ def plot_accuracy(x, results, title_suffix):
     plt.grid(True)
 
     plt.subplot(2, 2, 3)
-    plt.plot(x, results["cdf_abs_diff"], label="CDF Absolute Diff", marker='.')
+    plt.plot(x, results["cdf_abs_diff"], label="CDF Absolute Diff", marker=".")
     plt.xscale("log")
     plt.yscale("log")
     plt.xlabel("x")
@@ -91,7 +91,7 @@ def plot_accuracy(x, results, title_suffix):
 
[…]
