+++ examples/gaussian.py
@@ -20,6 +20,7 @@ guess = [(10, -18, 1), (4, -5.5, 10), (5, -1, 0.5), (10, 3, 1), (4, 15, 3)]
 fitter.fit(guess)
 
 f, ax = plt.subplots(1, 1, figsize=(12, 6))
-plotter = fitter.plot_fit(show_individuals=True, x_label='X_data', y_label='Y_data', title='XY_plot',
-                          data_label='XY_data', axis=ax)
+plotter = fitter.plot_fit(
+    show_individuals=True, x_label="X_data", y_label="Y_data", title="XY_plot", data_label="XY_data", axis=ax
+)
 plt.show()
