### 2025-10-26 ad861a7
[v110-minor-2-1]
1. beta test corrections in `utilities_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
