### 2025-10-27 32fcf61
[changelog1-minor-1-1]
1. Updated pull request links for previous releases in CHANGELOG.
