### 2025-11-01 fd2630c
[betaPrime-major-1]
0. added utils in `utilities_d.py` for beta prime distribution
1. added `betaPrime_d.py`
2. updated documentation with `betaPrime_d.rst`
3. updated examples with `betaPrime.py`
4. corrected an issue in notes for `chiSquare_d.py`
5. updated `environment[dev].yaml` and `requirements[dev].txt` so that sphinx build doesn't break.
6. updated `version.py` -> v1.0.8
7. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
