+++ src/pymultifit/distributions/betaPrime_d.py
@@ -1,18 +1,13 @@
-"""Created on Aug 14 00:45:37 2024"""
-
-from typing import Dict
-
-import numpy as np
-from scipy.special import betaincinv
+"""Created on Oct 31 18:28:47 2025"""
 
 from .backend import BaseDistribution
-from .utilities_d import beta_cdf_, beta_pdf_, beta_log_pdf_, beta_log_cdf_
-from .. import md_scipy_like, OneDArray
+from .utilities_d import beta_prime_pdf_, beta_prime_log_pdf_, beta_prime_cdf_, beta_prime_log_cdf_
+from .. import OneDArray
 
 
-class BetaDistribution(BaseDistribution):
+class BetaPrimeDistribution(BaseDistribution):
     r"""
-    Class for Beta distribution.
+    Class for Beta Prime distribution.
 
     :param amplitude: The amplitude of the PDF. Defaults to 1.0. Ignored if ``normalize`` is ``True``.
     :type amplitude: float, optional
@@ -32,23 +27,19 @@ class BetaDistribution(BaseDistribution):
     :param normalize: bool, optional If ``True``, the distribution is normalized so that the total area under the PDF equals 1. Defaults to ``False``.
     :type normalize: bool, optional
 
-    :raise NegativeAmplitudeError: If the provided value of amplitude is negative.
-    :raise NegativeAlphaError: If the provided value of :math:`\alpha` is negative.
-    :raise NegativeBetaError: If the provided value of :math:`\beta` is negative.
-
     Examples
     --------
     Importing libraries
 
-    .. literalinclude:: ../../../examples/basic/beta1.py
+    .. literalinclude:: ../../../examples/basic/betaPrime.py
        :language: python
        :linenos:
        :lineno-start: 3
        :lines: 3-7
 
-    Generating a standard :math:`\text{Beta}(2, 30)` distribution with ``pyMultiFit`` and ``scipy``.
+    Generating a standard :math:`\text{BetaPrime}(2, 30)` distribution with ``pyMultiFit`` and ``scipy``.
 
-    .. literalinclude:: ../../../examples/basic/beta1.py
+    .. literalinclude:: ../../../examples/basic/betaPrime.py
        :language: python
        :linenos:
        :lineno-start: 9
@@ -56,32 +47,34 @@ class BetaDistribution(BaseDistribution):
 
     Plotting **PDF** and **CDF**
 
-    .. literalinclude:: ../../../examples/basic/beta1.py
+    .. literalinclude:: ../../../examples/basic/betaPrime.py
        :language: python
        :linenos:
        :lineno-start: 14
        :lines: 14-29
 
-    .. image:: ../../../images/beta_example1.png
-       :alt: Beta distribution (5, 30)
+    .. image:: ../../../images/beta_prime1.png
+       :alt: BetaPrime distribution (5, 30)
        :align: center
 
-    Generating a shifted and translated :math:`\text{Beta}(2, 30)` distribution.
+    Generating a shifted and translated :math:`\text{BetaPrime}(2, 30, 5, 3)` distribution.
 
-    .. literalinclude:: ../../../examples/basic/beta2.py
+    .. literalinclude:: ../../../examples/basic/betaPrime.py
        :language: python
-       :lineno-start: 9
-       :lines: 9-12
+       :linenos:
+       :lineno-start: 33
+       :lines: 33-34
 
     Plotting **PDF** and **CDF**
 
-    .. literalinclude:: ../../../examples/basic/beta2.py
+    .. literalinclude:: ../../../examples/basic/betaPrime.py
        :language: python
-       :lineno-start: 14
-       :lines: 14-29
+       :linenos:
+       :lineno-start: 36
+       :lines: 36-51
 
[…]
