### 2025-11-01 16fec03
[betaPrime-minor-1]
1. added `test_betaPrime_d.py`
2. updated stats in `betaPrime_d.py`

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
