### 2025-11-01 6068d9c
[betaPrime-minor-1-1]
1. updated `test_beta_d.py` removing the constraints issues.

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
