+++ src/tests/test_distributions/test_beta_d.py
@@ -1,11 +1,9 @@
 """Created on Dec 14 06:55:53 2024"""
 
-import pytest
 from scipy.stats import beta
 
 from . import base_test_functions as btf
 from ...pymultifit.distributions import BetaDistribution
-from ...pymultifit.distributions.backend import errorHandling as erH
 
 
 class TestBetaDistribution:
@@ -20,21 +18,6 @@ class TestBetaDistribution:
 
         dist_normalized = BetaDistribution(amplitude=2.0, normalize=True)
         assert dist_normalized.amplitude == 1.0
-
-    @staticmethod
-    def test_constraints():
-        with pytest.raises(erH.NegativeAmplitudeError, match=f"Amplitude {erH.neg_message}"):
-            BetaDistribution(amplitude=-1.0, normalize=False)
-
-        # amplitude should be internally updated to 1.0 if `normalize` is called
-        distribution = BetaDistribution(amplitude=-1.0, normalize=True)
-        assert distribution.amplitude == 1.0
-
-        with pytest.raises(erH.NegativeAlphaError, match=f"Alpha {erH.neg_message}"):
-            BetaDistribution(alpha=-3.0)
-
-        with pytest.raises(erH.NegativeBetaError, match=f"Beta {erH.neg_message}"):
-            BetaDistribution(beta=-3.0)
 
     @staticmethod
     def test_edge_cases():
