### 2025-11-01 6f2f97f
[betaPrime-minor-1-2]
1. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
