### 2025-11-02 648a855
[v110-minor-3]
1. moved constants to `__init__.py`
2. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
