+++ src/pymultifit/distributions/betaPrime_d.py
@@ -1,4 +1,5 @@
 """Created on Oct 31 18:28:47 2025"""
+
 from typing import Dict
 
 import numpy as np
@@ -173,8 +174,8 @@ class BetaPrimeDistribution(BaseDistribution):
         mean_ = (s * mean_) + _l
 
         num_ = a * (a + b - 1)
-        den_ = (b - 2) * (b - 1)**2
+        den_ = (b - 2) * (b - 1) ** 2
         variance_ = num_ / den_ if b > 2 else np.inf
         variance_ = variance_ * s**2
 
-        return {'mean': mean_, 'variance': variance_, 'std': np.sqrt(variance_)}
+        return {"mean": mean_, "variance": variance_, "std": np.sqrt(variance_)}
