+++ src/pymultifit/distributions/utilities_d.py
@@ -74,28 +74,20 @@ import numpy as np
 import scipy.special as ssp
 from custom_inherit import doc_inherit  # type: ignore
 
-from .. import doc_style, LOG, INF, OneDArray
-
-TWO = 2.0
-SQRT_TWO = np.sqrt(TWO)
-LOG_TWO = LOG(TWO)
-LOG_SQRT_TWO = ssp.xlogy(0.5, TWO)
-
-PI = np.pi
-SQRT_PI = np.sqrt(PI)
-LOG_PI = LOG(PI)
-LOG_SQRT_PI = ssp.xlogy(0.5, PI)
-
-TWO_PI = 2 * PI
-SQRT_TWO_PI = np.sqrt(TWO_PI)
-LOG_TWO_PI = LOG(TWO_PI)
-LOG_SQRT_TWO_PI = ssp.xlogy(0.5, TWO_PI)
-
-INV_PI = 1.0 / PI
-TWO_BY_PI = 2.0 * INV_PI
-SQRT_TWO_BY_PI = np.sqrt(TWO_BY_PI)
-LOG_TWO_BY_PI = LOG(TWO_BY_PI)
-LOG_SQRT_TWO_BY_PI = ssp.xlogy(0.5, TWO_BY_PI)
+from .. import (
+    doc_style,
+    LOG,
+    INF,
+    OneDArray,
+    SQRT_TWO,
+    LOG_TWO,
+    PI,
+    SQRT_TWO_PI,
+    LOG_SQRT_TWO_PI,
+    TWO_BY_PI,
+    SQRT_TWO_BY_PI,
+    LOG_SQRT_TWO_BY_PI,
+)
 
 
 def suppress_numpy_warnings():
