+++ src/pymultifit/__init__.py
@@ -3,6 +3,7 @@
 from typing import Union, Tuple, List, Annotated
 
 import numpy as np
+import scipy.special as ssp
 from deprecated.sphinx import deprecated
 from numpy.typing import NDArray
 
@@ -42,6 +43,27 @@ LOG = np.log
 # taken from https://stackoverflow.com/a/19141711
 EPSILON = np.finfo(float).eps
 epsilon = np.sqrt(EPSILON)
+
+TWO = 2.0
+SQRT_TWO = np.sqrt(TWO)
+LOG_TWO = LOG(TWO)
+LOG_SQRT_TWO = ssp.xlogy(0.5, TWO)
+
+PI = np.pi
+SQRT_PI = np.sqrt(PI)
+LOG_PI = LOG(PI)
+LOG_SQRT_PI = ssp.xlogy(0.5, PI)
+
+TWO_PI = 2 * PI
+SQRT_TWO_PI = np.sqrt(TWO_PI)
+LOG_TWO_PI = LOG(TWO_PI)
+LOG_SQRT_TWO_PI = ssp.xlogy(0.5, TWO_PI)
+
+INV_PI = 1.0 / PI
+TWO_BY_PI = 2.0 * INV_PI
+SQRT_TWO_BY_PI = np.sqrt(TWO_BY_PI)
+LOG_TWO_BY_PI = LOG(TWO_BY_PI)
+LOG_SQRT_TWO_BY_PI = ssp.xlogy(0.5, TWO_BY_PI)
 
 ListOrNdArray = Union[List[int | float], np.ndarray]
 OneDArray = Annotated[NDArray[np.float64], "1D array"]
