### 2025-11-02 ec695f9
[changelog1-minor-2]
1. added release v1.0.8

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
