+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 This format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [v1.0.8] - 02/11/2025
+**Release:** [v1.0.8](https://pypi.org/project/pymultifit/1.0.8)  
+**Pull Request:** https://github.com/syedalimohsinbukhari/pyMultiFit/pull/111  
+**Issues:**
+- BetaPrime (https://github.com/syedalimohsinbukhari/pyMultiFit/issues/109)  
+
+#### ADDED
+- `BetaPrimeDistribution`
+- `betaPrime.py` in examples
+- updated sphinx documentation.
+- added tests to `test_BetaPrimeDistribution`.
+
+---
 ## [v0.1.2] — 19/08/2024
 **Release:** [v0.1.2](https://pypi.org/project/pymultifit/0.1.2)  
 **Pull Reqeust:** https://github.com/syedalimohsinbukhari/pyMultiFit/pull/5  
@@ -14,6 +27,7 @@ This format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 - Generator functions (https://github.com/syedalimohsinbukhari/pyMultiFit/issues/9)
 
 #### OVERVIEW
+
 ---
 ## [v0.1.1] — 03/08/2024
 **Release:** [v0.1.1](https://pypi.org/project/pymultifit/0.1.1)  
@@ -21,8 +35,7 @@ This format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and
 **Issues:**
 - Parameter extraction (https://github.com/syedalimohsinbukhari/pyMultiFit/issues/3)
 
-#### OVERVIEW
-This release introduces minor code base enhancements.
+Introduces minor code base enhancements.
 
 #### ADDED
 ##### Core framework
@@ -35,8 +48,7 @@ This release introduces minor code base enhancements.
 **Pull Request:** https://github.com/syedalimohsinbukhari/pyMultiFit/pull/1  
 **Issues:**
 
-### Overview
-This release introduces the **first stable version** of `pyMultiFit`, establishing the base framework for multi-distribution fitting and model generation.
+**First stable version** of `pyMultiFit`, establishing the base framework for multi-distribution fitting and model generation.
 
 ### Added
 
