### 2025-11-03 0bed6de
[johnsonsu-minor-1]
1. added `johnsonSU_d.py` to distributions and tests
2. updated `base_test_functions.py` with two separate shape parameters
3. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
