+++ src/tests/test_distributions/test_logNormal_d.py
@@ -45,7 +45,7 @@ class TestLogNormalDistribution:
         btf.stats(
             custom_distribution=LogNormalDistribution.from_scipy_params,
             scipy_distribution=lognorm,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             median=False,
         )
 
@@ -54,7 +54,7 @@ class TestLogNormalDistribution:
         btf.value_functions(
             custom_distribution=LogNormalDistribution.from_scipy_params,
             scipy_distribution=lognorm,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             log_check=True,
         )
 
@@ -63,6 +63,6 @@ class TestLogNormalDistribution:
         btf.single_input_n_variables(
             custom_distribution=LogNormalDistribution.from_scipy_params,
             scipy_distribution=lognorm,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             log_check=True,
         )
