+++ src/tests/test_distributions/test_chiSquare_d.py
@@ -38,7 +38,7 @@ class TestChiSquareDistribution:
         btf.stats(
             custom_distribution=ChiSquareDistribution.from_scipy_params,
             scipy_distribution=chi2,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             median=False,
         )
 
@@ -47,7 +47,7 @@ class TestChiSquareDistribution:
         btf.value_functions(
             custom_distribution=ChiSquareDistribution.from_scipy_params,
             scipy_distribution=chi2,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             log_check=True,
         )
 
@@ -56,6 +56,6 @@ class TestChiSquareDistribution:
         btf.single_input_n_variables(
             custom_distribution=ChiSquareDistribution.from_scipy_params,
             scipy_distribution=chi2,
-            parameters=[btf.shape_parameter, btf.loc_parameter, btf.scale_parameter],
+            parameters=[btf.shape1_parameter, btf.loc_parameter, btf.scale_parameter],
             log_check=True,
         )
