### 2025-11-03 a30adc5
[johnsonsu-minor-2]
1. added another location parameter to `base_test_functions.py`
2. minor test changes
