### 2025-11-03 725a3ed
[gumbel-minor-1]
1. added gumbel distribution in `utilities_d.py`
2. updated version to v1.0.10 in version.py`
3. added more decorators in `__init__.py`
4. minor changes

Signed-off-by: Syed Ali Mohsin Bukhari <syedali.b@outlook.com>
