+++ src/pymultifit/distributions/utilities_d.py
@@ -32,6 +32,10 @@ __all__ = [
     "gamma_log_pdf_",
     "gamma_cdf_",
     "gamma_log_cdf_",
+    "gumbel_pdf_",
+    "gumbel_log_pdf_",
+    "gumbel_cdf_",
+    "gumbel_log_cdf_",
     "sym_gen_normal_pdf_",
     "sym_gen_normal_cdf_",
     "sym_gen_normal_log_pdf_",
@@ -71,7 +75,6 @@ __all__ = [
     "cubic",
 ]
 
-import functools
 from typing import Union, Callable
 
 import numpy as np
@@ -79,40 +82,23 @@ import scipy.special as ssp
 from custom_inherit import doc_inherit  # type: ignore
 
 from .. import (
+    check_scale_positive,
     doc_style,
     LOG,
     INF,
     OneDArray,
     SQRT_TWO,
     LOG_TWO,
-    PI,
     SQRT_TWO_PI,
     LOG_SQRT_TWO_PI,
     TWO_BY_PI,
     SQRT_TWO_BY_PI,
     LOG_SQRT_TWO_BY_PI,
+    suppress_numpy_warnings,
 )
 
 
-def suppress_numpy_warnings():
-    """
-    A decorator that suppresses NumPy warnings using np.errstate.
-
-    Parameters (all optional):
-        divide, over, under, invalid: Can be 'ignore', 'warn', 'raise', 'call', 'print', or 'log'
-    """
-
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            with np.errstate(all="ignore"):
-                return func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
-
-
+@check_scale_positive
 @suppress_numpy_warnings()
 def arc_sine_pdf_(
     x: OneDArray, amplitude: float = 1.0, loc: float = 0.0, scale: float = 1.0, normalize: bool = False
@@ -159,9 +145,6 @@ def arc_sine_pdf_(
     if y.size == 0:
         return y
 
-    if scale < 0:
-        return np.full(y.shape, np.nan)
-
     c1 = (y > 0) & (y < 1)
     c2 = y == 0
     c3 = y == 1
@@ -177,6 +160,7 @@ def arc_sine_pdf_(
     return pdf_
 
 
+@check_scale_positive
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_log_pdf_(
@@ -217,6 +201,7 @@ def arc_sine_log_pdf_(
     return log_pdf_
 
 
+@check_scale_positive
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_pdf_, style=doc_style)
 def arc_sine_cdf_(
@@ -227,15 +212,13 @@ def arc_sine_cdf_(
     if y.size == 0:
         return y
 
-    if scale < 0:
-        return np.full(y.shape, np.nan)
-
     c1 = (y > 0) & (y < 1)
     c2 = y < 1
 
     return np.select(condlist=[c1, c2], choicelist=[TWO_BY_PI * np.arcsin(np.sqrt(y)), 0.0], default=1.0)
 
 
+@check_scale_positive
 @suppress_numpy_warnings()
 @doc_inherit(parent=arc_sine_cdf_, style=doc_style)
 def arc_sine_log_cdf_(
@@ -267,6 +250,7 @@ def arc_sine_log_cdf_(
[…]
